\name{add.array.values}
\alias{add.array.values}
\title{aggiungi array di una nuova variabile all'oggetto values}
\usage{
add.array.values(values, new.array, variable = "New Data")
}
\description{
aggiungi array di una nuova variabile all'oggetto values
}
\seealso{
Other Array Values: \code{\link{get.array.values}};
\code{\link{get.origin}}; \code{\link{set.origin}}
}

