% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{add.colours.contours}
\alias{add.colours.contours}
\title{Add color definitions to contours}
\usage{
add.colours.contours(contours)
}
\arguments{
\item{contours}{the contours dataframe.}
}
\value{
contours dataframe with colour definition for each contour.
}
\description{
Colours will be added using the "raimbow" colour selection and will be stored in a new "display.colour" column of the dataframe.
}
\seealso{
Other Contours: \code{\link{display.contours.legend}},
  \code{\link{get.contours}},
  \code{\link{sanitize.contours}},
  \code{\link{volume.contours}}
}
\concept{Contours}
