% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{add.vois}
\alias{add.vois}
\title{Combine 2 existing VOIS object into one}
\usage{
add.vois(vois, vois.new)
}
\arguments{
\item{vois, vois.new}{the VOIS objects to be combined.}
}
\value{
a VOIS object.
}
\description{
Combine 2 existing VOIS object into one
}
\seealso{
Other VOIs: \code{\link{create.vois.from.logical.array}},
  \code{\link{create.vois}}, \code{\link{get.subset.vois}},
  \code{\link{get.voi.logical}},
  \code{\link{get.voiindex}}, \code{\link{get.vois}},
  \code{\link{is.in.voi}}
}

