% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{alpha.beta.ion.range}
\alias{alpha.beta.ion.range}
\title{Evaluation of ion alpha and beta LQ parameters}
\usage{
alpha.beta.ion.range(model = "MKM", model.parameters = data.frame(alpha0 =
  0.1295, beta0 = 0.03085, rN = 4, rd = 0.31), cell.name = NULL,
  particle = "H", energies = NULL, lets = NULL,
  calculusType = "rapidMKM", precision = 0.5, ignore.stdout = TRUE,
  ignore.stderr = TRUE, get.raw.data = FALSE, remove.temp.files = TRUE)
}
\arguments{
\item{model}{the name of the model (options: 'MKM', 'LEMI', 'LEMII', 'LEMIII')}

\item{model.parameters}{a dataframe containing the "biological" parameter associated to a specific biological tissue/cell line. In the case of the MKM:
\itemize{
  \item{alpha0} LQ alpha parameter of the reference radiation [Gy^-1]
  \item{beta0} LQ beta parameter of the reference radiation [Gy^-2]
  \item{rN} cell nucleus radius [um]
  \item{rd} domain radius [um]
}
In the case of LEMI, LEMII, LEMIII:
\itemize{
  \item{alpha0} LQ alpha parameter of the reference radiation [Gy^-1]
  \item{beta0} LQ beta parameter of the reference radiation [Gy^-2]
  \item{rN} cell nucleus radius [um]
  \item{Dt} threshold dose [Gy]
}
Optionally, instead of the specific values, the range of the parameters can be specified in the same data.frame. In the case of MKM:
\itemize{
  \item{alpha0.min, alpha0.max, alpha0.N} range of variability for parameter \code{alpha0} and number of steps
  \item{beta0.min, beta0.max, beta0.N} range of variability for parameter \code{beta0} and number of steps
  \item{rN.min, rN.max, rN.N} range of variability for parameter \code{rN} and number of steps
  \item{rd.min, rN.max} domain radius [um]
}}

\item{particle}{available particles: 'H', 'He', 'Li', 'Be', 'B,', 'C', 'N', 'O', 'F', 'Ne'.}

\item{energies}{vector of energies for the particle [MeV]}

\item{lets}{vector of LETs for the particle [keV/um]. It is used if \code{energies} is \code{NULL}.}

\item{calculusType}{the type of the evaluations options are:
\itemize{
  \item{rapidMKM} fast implementation for the MKM (Kase2008)
  \item{rapidScholz} fast implementation of the LEMI, LEMII, LEMIII (Scholz2000)
  \item{rapidRusso} fast implementation of the LEMI, LEMII, LEMIII, more accurate than the \code{rapidScholz} (Russo2010)
  \item{slow_alphaIon_betaIon} slow Monte Carlo evaluation (compatible with MKM, LEMI, LEMII and LEMIII).
}}

\item{precision}{used only for the Monte Carlo evaluation. If precision < 1.0 it represents the target relative standard deviation to be reached in the evaluations of the cell survivals. If precision >= 1 it represents the number of cells to be simulated.}

\item{get.raw.data}{If calculusType=slow_alphaIon_betaIon, returns the raw data (survival vs dose for each simulated cell) from which alpha and beta parameters can be extracted, otherwhise it is ignored.}

\item{cellType}{name of the tissue/cell line (optional)}
}
\value{
a data.frame containing all the information specified including the alpha and beta MKM evaluation (note, in the MKM implementation beta = betaX).
}
\description{
The evaluation is performed for a monoenergetic ion. It uses a C++ implementation for the evaluation. It accepts a single set of parameter depending on the choosen radiobiological model, or alternatively a full range of variability (min,max) for each parameter.
}
\seealso{
Other LEM/MKM Models: \code{\link{alpha.beta.mkm}},
  \code{\link{alpha.fun.mkm}}
}
