% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{alpha.beta.ion.sequence}
\alias{alpha.beta.ion.sequence}
\title{Evaluation of ion alpha and beta LQ parameters}
\usage{
alpha.beta.ion.sequence(model = "MKM", parameters = data.table(cell =
  "test", alpha0 = 0.1295, beta0 = 0.03085, rN = 4, rd = 0.31, particle = "H",
  energy = 50, let = NA), calculusType = "rapidMKM", precision = 0.5,
  ignore.stdout = TRUE, ignore.stderr = TRUE, remove.temp.files = TRUE)
}
\arguments{
\item{model}{the name of the model (options: 'MKM', 'LEMI', 'LEMII', 'LEMIII')}

\item{parameters}{input data.frame containing the cell name (column 'cell', optional) "biological" parameters associated to a specific biological tissue/model, and the specification of the primary particles (ion type and energy and/or LET). All the parameters have to be put in a dataframe. Each row of the dataframe represent the irradiation of a specific tissue/model with a specific particle. In the case of the MKM the biological parameters
\itemize{
  \item{alpha0} LQ alpha parameter of the reference radiation [Gy^-1]
  \item{beta0} LQ beta parameter of the reference radiation [Gy^-2]
  \item{rN} cell nucleus radius [um]
  \item{rd} domain radius [um]
}
In the case of LEMI, LEMII, LEMIII:
\itemize{
  \item{alpha0} LQ alpha parameter of the reference radiation [Gy^-1]
  \item{beta0} LQ beta parameter of the reference radiation [Gy^-2]
  \item{rN} cell nucleus radius [um]
  \item{Dt} threshold dose [Gy]
}
The ion is specified with the columns:
\itemize{
  \item{particle} ion type: 'H', 'He', 'Li', 'Be', 'B', 'C', 'N', 'O', 'F' or 'Ne'
  \item{energy} Energy of the particle in MeV/u
  \item{let} LET of the particle in keV/um. It is used only if the energy is not specified.
}
#'}
}
\description{
The evaluation is performed for a monoenergetic ion. It uses a C++ implementation for the evaluation. It accepts a sequence of parameters and ion specifications stored in an input data.frame.
}
\seealso{
Other LEM/MKM Models: \code{\link{alpha.beta.ion.range}};
  \code{\link{alpha.beta.mkm}}; \code{\link{alpha.fun.mkm}}
}

