% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{alpha.beta.lem}
\alias{alpha.beta.lem}
\title{Evaluation of alpha and beta LQ parameter (LEM)}
\usage{
alpha.beta.lem(alphaX = 0.1, betaX = 0.05, rN = 5, Dt = 30,
  alphaX.min = NULL, alphaX.max = NULL, alphaX.N = NULL,
  betaX.min = NULL, betaX.max = NULL, betaX.N = NULL, rN.min = NULL,
  rN.max = NULL, rN.N = NULL, Dt.min = NULL, Dt.max = NULL,
  Dt.N = NULL, model = "LEMI", calculusType = "rapidScholz",
  cellType = NULL, particleType = "H", energies = NULL, lets = NULL,
  ignore.stdout = TRUE, ignore.stderr = TRUE)
}
\arguments{
\item{alphaX,betaX,rN,Dt}{LEM "biological" parameter associated to a specific biological tissue/cell line:
\itemize{
  \item{alphaX} LQ alpha parameter of the reference radiation [Gy^-1]
  \item{betaX} LQ beta parameter of the reference radiation [Gy^-2]
  \item{rN} cell nucleus radius [um]
  \item{Dt} threshold dose [Gy]
}}

\item{alphaX.min,alphaX.max}{range of variability for parameter \code{alphaX}}

\item{alphaX.N}{number of step for \code{alphaX}}

\item{betaX.min,betaX.max}{range of variability for parameter \code{betaX}}

\item{betaX.N}{number of step for \code{betaX}}

\item{rN.min,rN.max}{range of variability for parameter \code{rN}}

\item{rN.N}{number of step for \code{rN}}

\item{Dt.min,Dt.max}{range of variability for parameter \code{Dt}}

\item{Dt.N}{number of step for \code{Dt}}

\item{model}{The specific LEM model (available: 'LEMI', 'LEMII', 'LEMIII')}

\item{calculusType}{Approximation used for the evaluation (available: 'rapidSholz', 'rapidRusso')}

\item{cellType}{name of the tissue/cell line (optional)}

\item{energies}{vector of energies for the particle [MeV]}

\item{lets}{vector of LETs for the particle [keV/um]. It is used if \code{energies} is \code{NULL}.}

\item{particel}{type. Available ions: 'H', 'He', 'Li', 'Be', 'B,', 'C', 'N', 'O', 'F', 'Ne'.}
}
\value{
a data.frame containing all the information specified including the alpha and beta LEM evaluation.
}
\description{
The evaluation is performed for a monoenergetic ion. It uses a C++ LEM "rapid" implementation for the evaluation. It accepts a single set of LEM parameter (\code{alphaX}, \code{betaX}, \code{rN}, \code{Dt}), or alternatively a full range of variability (min,max) for each parameter.
}
\seealso{
Other LEM/MKM Models: \code{\link{alpha.beta.mkm}};
  \code{\link{alpha.fun.mkm}}
}

