% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{alpha.beta.lqp}
\alias{alpha.beta.lqp}
\title{Evaluation of Linear-Quadratic (LQ) Poisson Model parameters}
\usage{
alpha.beta.lqp(R = R, D50 = D50, gam = gam, d.ref = 2, d = 2)
}
\arguments{
\item{R}{the ratio alphaX/betaX (Gy).}

\item{D50}{the tolerance dose at 50\%.}

\item{gam}{the gamma parameter (associated to the clonogenic density).}

\item{d.ref}{the reference dose at which D50 and gam are fitted (usually 2 Gy).}

\item{d}{the actual level of dose for the evaluation of alphaX, betaX.}
}
\value{
A list consisting of:
\item{alphaX}{the LQ alpha parameter}
\item{betaX}{the LQ beta parameter}
}
\description{
It translates R, D50 and gamma parameters to alphaX and betaX.
}
\details{
Note: d.ref è la dose per frazione di riferimento
in cui sono fittati i dati per D50, gam.
Per il momento d.ref non è usato: si assume una condizione di riferimento
standard per D50 e gam (forse... )
}
\seealso{
Other Basic Biological Models: \code{\link{rbe.evaluate}}
}

