% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alpha.fun.mkm}
\alias{alpha.fun.mkm}
\title{Evaluate alpha function (MKM)}
\usage{
alpha.fun.mkm(alphaX = 0.1295, betaX = 0.03085, rN = 4, rd = 0.31,
  cellType = NULL, particleType = "H", energies = NULL, lets = NULL,
  use.limits = FALSE, outmessages = FALSE)
}
\arguments{
\item{alphaX,betaX,rN,rd}{The radiobiological parameter of the MKM.}

\item{cellType}{Optionally you can associate a cell name tag (I have to figure yet what use to do with that).}

\item{particleType}{The particle specification ('H', 'He', 'Li', 'Be', 'B', 'C', 'N', 'O', 'F')}

\item{energies,lets}{A vector of specific energies (MeV/u) or LETs (keV/um).}

\item{use.limits}{Filters the energies or the LETs values to rule out values too high or too low.}
}
\description{
Returns a function alpha(LET) or alpha(specific energy), using the MKM model. The functions is an interpolation functions that uses a set of alpha values evaluated over the defined set of LETs (or specific energies).
}
\seealso{
Other LEM/MKM Models: \code{\link{alpha.beta.lem}};
  \code{\link{alpha.beta.mkm}}
}

