% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gate-functions.R
\name{available.variables.gate}
\alias{available.variables.gate}
\title{Available variables (Gate)}
\usage{
available.variables.gate()
}
\value{
A list of the available variables
}
\description{
The available variables name identifiers that can be evaluated with Gate.
}

