\name{available.variables}
\alias{available.variables}
\title{Available variables (PlanKIT)}
\usage{
available.variables()
}
\value{
A list of the available variables
}
\description{
The available variables name identifiers that can be
evaluated with PlanKIT
}
\seealso{
Other PlanKITVariables: \code{\link{available.beamlines}};
\code{\link{available.constraint.types}}
}

