% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{check.gate}
\alias{check.gate}
\title{Check the configuration of Gate}
\usage{
check.gate(use.warning = FALSE)
}
\description{
If the check fails it generates an error.
#'@param use.warning if TRUE it uses warning instead of an error.
}
\seealso{
Other Environment: \code{\link{check.plankit}},
  \code{\link{get.dektools.env}},
  \code{\link{set.dektools.env}},
  \code{\link{set.lut.folder}},
  \code{\link{setenv.rplanit}}
}

