% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{combine.beamLUTs}
\alias{combine.beamLUTs}
\title{Combine beamLUTs}
\usage{
combine.beamLUTs(beamLUTs.list, vois, contours)
}
\arguments{
\item{beamLUTs.list}{the list of the beamLUTs. The elements of this list have to be named according to the names of the tissues specified in the contours (e.g. list(tissue1=beamLUTs1, tissue2=beamLUTs2, ....))}

\item{vois}{the vois object.}

\item{contours}{the contours object.}
}
\value{
a beamLUTs object.
}
\description{
Combine a list of beamLUTs in a unique beamLUTs object. The voxel values are taken from the different beamLUTs in the list, according to the tissue
specification in the contours and their belonging to the corresponding VOI. In case of a voxel belonging to multiple VOIS, the last tissue specification in the contours will overwrite the previous ones. The beamLUTs in the list and the vois object should have the same dimensionality.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}},
  \code{\link{get.bdose.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.letd.beamLUT}},
  \code{\link{get.rbe.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{read.beamLUTs}}, \code{\link{read.beamLUT}},
  \code{\link{write.beamLUT}}
}
