% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{combine.values}
\alias{combine.values}
\title{Combine values}
\usage{
combine.values(values.list, vois, contours)
}
\arguments{
\item{values.list}{the list of the values. The elements of this list have to be named according to the names of the tissues specified in the contours (e.g. list(tissue1=values1, tissue2=values2, ....))}

\item{vois}{the vois object.}

\item{contours}{the contours object.}
}
\value{
a values object.
}
\description{
Combine a list of values in a unique values object. The voxel values are taken from the different values in the list, according to the tissue
specification in the contours and their belonging to the corresponding VOI. In case of a voxel belonging to multiple VOIS, the last tissue specification in the contours will overwrite the previous ones. The values in the list and the vois object should have the same dimensionality.
}
\seealso{
Other Values: \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{merge.values}},
  \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{threshold.values}},
  \code{\link{values.from.sparse.array}}
}
\concept{Values}
