% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beams-functions.R
\name{convert.fluence2mu}
\alias{convert.fluence2mu}
\title{converte fluenze in beams da numero di particelle a MU}
\usage{
convert.fluence2mu(beams, type = "protonCnao")
}
\description{
secondo le curve di calibrazione specificate
}
\seealso{
Other Beams: \code{\link{add.field}},
  \code{\link{convert.mu2fluence}},
  \code{\link{create.beam}}, \code{\link{create.field}},
  \code{\link{display.beamports}},
  \code{\link{display.beams}}, \code{\link{get.beams}},
  \code{\link{get.ct.entrance.point}},
  \code{\link{get.isocenter}}, \code{\link{get.rays}},
  \code{\link{read.beams}}, \code{\link{write.beams}}
}
\concept{Beams}
