% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.mu2fluence}
\alias{convert.mu2fluence}
\title{converte fluenze in beams da MU a numero di particelle}
\usage{
convert.mu2fluence(beams, type = "protonCnao")
}
\description{
secondo le curve di calibrazione specificate
}
\seealso{
Other Beams: \code{\link{add.field}};
  \code{\link{convert.fluence2mu}};
  \code{\link{create.beam}}; \code{\link{create.field}};
  \code{\link{display.beamports}};
  \code{\link{display.beams}}; \code{\link{get.beams}};
  \code{\link{get.ct.entrance.point}};
  \code{\link{get.isocenter}}; \code{\link{get.rays}};
  \code{\link{read.beams.fluka}}; \code{\link{read.beams}};
  \code{\link{write.beams}}
}

