\name{create.beam}
\alias{create.beam}
\title{Create an empty beams object}
\usage{
create.beam(nbeams = 1, with.spots = FALSE)
}
\arguments{
  \item{nbeams}{The number of the beams}

  \item{with.spots}{Include spot coordinates}
}
\description{
Create an empty beams object
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.field}};
\code{\link{display.beamports}};
\code{\link{display.beams}}; \code{\link{get.beams}};
\code{\link{get.ct.entrance.point}};
\code{\link{get.isocenter}}; \code{\link{get.rays}};
\code{\link{read.beams.fluka}}; \code{\link{read.beams}};
\code{\link{write.beams}}
}

