\name{create.gate.structure}
\alias{create.gate.structure}
\title{Create Gate structure}
\usage{
create.gate.structure(plan)
}
\arguments{
  \item{plan}{A Plan Object.}
}
\description{
Create the folder structure of a Gate simulation for a
specific plan.
}
\seealso{
Other PlanGate: \code{\link{create.plan.gate}};
\code{\link{create.waterbox.gate}};
\code{\link{file.variable.gate}}; \code{\link{set.ct.gate}}
}

