% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.intervals}
\alias{create.intervals}
\title{Generate bin intervals from points}
\usage{
create.intervals(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector (with N+1 elements)
}
\description{
Generate bin intervals from points
}
\seealso{
Other Values, Values Utilities: \code{\link{generate.values.from.events}};
  \code{\link{get.coordinates}};
  \code{\link{get.voxel.index}}
}

