% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noise-functions.R
\name{create.noise.ct}
\alias{create.noise.ct}
\title{noise sulla CT}
\usage{
create.noise.ct(ct, lambda = 0.15, sigma = 0.025, sigmaHU = NULL,
  WGN = FALSE, return.noise = FALSE)
}
\description{
noise sulla CT
}

