% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plankit-functions.R
\name{create.plan}
\alias{create.plan}
\title{Generate Plan template (Pure-dek)}
\usage{
create.plan(plan = NA, name = "test", ctFile = "ct.3d", ct = NULL,
  contoursFile = "plan.contours", contours = NULL,
  hounsfieldToDensityFile = NULL, hounsfieldToStoichiometryFile = NULL,
  computingGridVoxelSizes = c(2, 2, 2), computingGridCoverage = 0,
  computingGridCoverageBoundary = 0, beamLines = NULL,
  computingValues = c("Dose[Gy]"), cutOffRadius = 100,
  cutEnergyFraction = 0.01, fields = NULL, prescription = NULL,
  maxOptimizationIterations = 50, outputBeamsFile = NULL,
  inputBeamsFile = NULL, saveBeamLUTs = FALSE, outputBeamLUTFile = NULL,
  inputBeamLUTFile = NULL, vois = NULL, outputVoisFile = NULL,
  inputVoisFile = NULL, values = NULL, outputValuesFile = NULL,
  inputValuesFile = NULL, beams = NULL)
}
\value{
A Plan (plankit.plan object).
}
\description{
Generate a default (empty) template of the Plan to be used with PlanKIT.
}
\seealso{
Other Plan: \code{\link{read.plan}};
  \code{\link{save.plan}}
}

