% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{create.vois.from.logical.array}
\alias{create.vois.from.logical.array}
\title{Create a VOIS object from a logical array}
\usage{
create.vois.from.logical.array(logical.array, x, y, z, voi)
}
\arguments{
\item{logical.array}{The logical array (TRUE -> the voi is present in the voxel.}

\item{x, y, z}{The coordinates of the x,y,z axes for the VOIS 3D grid.}

\item{voi}{the VOI name.}
}
\value{
a VOIS object.
}
\description{
Create a VOIS object from a logical array
}
\seealso{
Other VOIs: \code{\link{add.vois}},
  \code{\link{create.vois}}, \code{\link{get.subset.vois}},
  \code{\link{get.voi.logical}},
  \code{\link{get.voiindex}}, \code{\link{get.vois}},
  \code{\link{is.in.voi}}
}

