% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{dataframe.from.values}
\alias{dataframe.from.values}
\title{Generate data.frame from values object}
\usage{
dataframe.from.values(values = values, vois = vois, variables = NULL,
  rois = NULL)
}
\description{
Uses only voxels belogning to the specified VOIs.
Each row of the output data.frame is a single voxel.
Each voxel is associated to a VOI.
Voxels assiciated to different VOIs are replicated on different rows.
}
\seealso{
Other Values: \code{\link{combine.values}},
  \code{\link{create.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{values.from.sparse.array}}
}

