% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-functions.R
\name{demo.puredek}
\alias{demo.puredek}
\title{Evaluate a test planning}
\usage{
demo.puredek(plan.name = "test-plan", particle = "proton", dose = 1,
  outmessages = FALSE)
}
\arguments{
\item{plan.name}{The name of the plan.}

\item{particle}{The primary particle ('proton', 'lithium', 'helium', 'carbon' or 'oxygen')}

\item{dose}{The equivalent dose to be delivered in the PTV (Gy(RBE))}

\item{outmessages}{Display detailed output messages of the evaluation. Warning: the throughput is high, use this option only when running from a simple text terminal.}
}
\value{
The plan object.
}
\description{
Generate a simple water phantom CT with a central PTV and four OARs around.
Set a plan with protons (default) with 4 coplanar fields and perform an optimization to deliver exatly 1 Gy(RBE) (default) in the PTV.
Generate a graphical report.
}

