% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.R
\name{display.all.plan}
\alias{display.all.plan}
\title{procedura per creare un display complessivo da un "plan"}
\usage{
display.all.plan(plan)
}
\description{
procedura per creare un display complessivo da un "plan"
}
