\name{display.beams}
\alias{display.beams}
\title{Beam statistics plot}
\usage{
display.beams(beams, plan = NULL)
}
\arguments{
  \item{beams}{beams dataframe}
}
\description{
Beam statistics plot
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.beam}}; \code{\link{create.field}};
\code{\link{get.beams}}; \code{\link{get.isocenter}};
\code{\link{read.beams}}; \code{\link{write.beams}}
}

