% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display-functions.R
\name{display.dvh.old}
\alias{display.dvh.old}
\title{plot del dvh per una variabile specifica}
\usage{
display.dvh.old(dvh, plan = NULL, Diff = FALSE)
}
\description{
plot del dvh per una variabile specifica
}

