% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.R
\name{display.dvh}
\alias{display.dvh}
\title{Display DVHs}
\usage{
display.dvh(dvh, plan = NULL, Diff = FALSE, alpha.color = 1,
  title = NULL, show.plot = TRUE, decimate = TRUE,
  show.prescription = FALSE, file.name = NULL, height = 7, width = 7,
  fixed.scale = FALSE, return.dataframe = FALSE)
}
\description{
Display a single DVH or a list of DVHs
}
\details{
@param dvh single DVH (created via \code{\link{dvh.evaluate}}) or a list of DVHs
 @param plan plan object (optional)
 @param Diff display "differential" DVH (boolean, optional)
 @param alpha.color opacity of the plot (optional)
 @param title the title of the plot (optional)
 @param show.plot if \code{TRUE} it display the plot on screen. If \code{FALSE} it returns a ggplot plot structure if (optional)
 @param decimate if the DVH points are too much (> 1000), it plots only a sample of them, to speed up the visualization (boolean, optional)
 @param show.prescription plot the prescription over the DVH. The plan object needs to be specified also (boolean, optional)
 @param filename the name of the file in which to save the figure (optional)
 @param height,weight sizes (inches) of the figure to be saved in the file (optional)
 @param fixed.scale Forces a common fixed scale when faceting (if there are different variables).
 @param return.dataframe Return the data.frame of the DVH(s).

@return If \code{show.plot} is \code{FALSE}, it returns a ggplot2 plot structure to be used for further processing. If \code{return.dataframe = TRUE} it returns the generated data.frame of the DVH(s).

@family display dvh
 @export
 @import ggplot2
}

