% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.R
\name{display.profiles}
\alias{display.profiles}
\title{Display profiles}
\usage{
display.profiles(profile.values, profile.ct = NULL, show.ct.legend = TRUE,
  y.axis.name = NULL, x.axis.name = NULL, name = NULL, depth.lim = NULL,
  file.name = NULL, height = 7, width = 7, dpi = 300,
  show.plot = TRUE, return.plot = FALSE)
}
\arguments{
\item{profile.values}{The main profiles dataframe to be plotted. It can contain a single profile or a multiple of profiles.}

\item{profile.ct}{the Background ct (or other data) profile.}

\item{show.ct.legend}{Boolean. A colorbar for the levels containted in profile.ct is displayed if TRUE.}

\item{y.axis.name}{Optional label for the y axis (default uses the data stored in profile.values if unique).}

\item{x.axis.name}{Optional label for the x axis (default uses the data stored in profile.values if unique).}

\item{name}{Optional label for the set of profiles.}

\item{depth.lim}{Optional limits for the x axis.}

\item{file.name}{File name for saving the plot.}

\item{height}{The height of the saved plot image (inches).}

\item{width}{The width of the saved plot image (inches).}

\item{show.plot}{Show plot.}

\item{return.plot}{return a ggplot2 plot if TRUE.}
}
\description{
Display a single or multiple profiles. Optionally the profile could be plotted over a background color gradient representing the ct data (or other values).
}
\seealso{
Other Profiles: \code{\link{display.profile}}
}
