\name{display.rays}
\alias{display.rays}
\title{Display rays (3D)}
\usage{
display.rays(rays, alpha = 1, ray.length = 1, add = FALSE)
}
\arguments{
  \item{rays}{the rays dataframe.}

  \item{alpha}{the opacity value for the point and lines.}

  \item{ray.length}{Length of the rays. If ray.lengt=0 ti
  will plot only the spots.}

  \item{add}{Add to existing 3D plot.}
}
\description{
Display rays (3D)
}

