% Generated by roxygen2 (4.0.1): do not edit by hand
\name{display.slice.ct}
\alias{display.slice.ct}
\title{Display CT slices}
\usage{
display.slice.ct(ct, contours = NULL, x = NA, y = NA, z = NA, Nx = NA,
  Ny = NA, Nz = NA, plan = NULL, xlim = NULL, ylim = NULL,
  zlim = NULL, vlim = NULL, HU.window = c(-1000, 3000),
  invert.y.axis = FALSE, file.name = NULL, width = 7, height = 7)
}
\description{
Alias for display.slices().
}
\seealso{
Other display slices: \code{\link{display.slice.all}};
  \code{\link{display.slice}};
  \code{\link{dumpslices.3d.2.png}}
}

