\name{display.spots}
\alias{display.spots}
\title{Display spots (3D)}
\usage{
display.spots(beams, vois = NULL, voi = NULL, display.iso = TRUE,
  alpha.spot = 0.2, alpha.voi = 1)
}
\arguments{
  \item{beams}{the beams dataframe}

  \item{vois}{the vois object}

  \item{voi(s)}{names}

  \item{display.iso}{display the isocenter}

  \item{alpha.spot}{the opacity value for the spot points}
}
\description{
Display spots (3D)
}

