% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display-functions.R
\name{dumpslices.3d.2.png}
\alias{dumpslices.3d.2.png}
\title{fa un dump del file .3d in una serie di immagini png corrispondenti a ciascuna
slice in z.}
\usage{
dumpslices.3d.2.png(file.name, x.min = -Inf, x.max = +Inf, y.min = -Inf,
  y.max = +Inf, z.min = -Inf, z.max = +Inf)
}
\description{
fa un dump del file .3d in una serie di immagini png corrispondenti a ciascuna
slice in z.
}
\seealso{
Other display slices: \code{\link{display.slice.all}};
  \code{\link{display.slice.ct}};
  \code{\link{display.slice}}
}

