% Generated by roxygen2 (4.0.2): do not edit by hand
\name{embed.data}
\alias{embed.data}
\title{Embed data into plan}
\usage{
embed.data(plan, embed.ct = TRUE, embed.contours = TRUE,
  embed.beams = TRUE, embed.values = TRUE, embed.vois = TRUE)
}
\arguments{
\item{plan}{A plan.}

\item{embed.ct,embed.contours,embed.beams,embed.values,embed.vois}{By default all data are embedded.}
}
\value{
A plan with embedded data.
}
\description{
Embed the data (CT, contours, beams, values, vois) associated with a plan directly in the plan object (to provide portability).
}
\seealso{
Other PlanKIT: \code{\link{run.dek.forward}};
  \code{\link{run.dek.inverse}}
}

