% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{evaluate.epl}
\alias{evaluate.epl}
\title{Evaluate the equivalent path length}
\usage{
evaluate.epl(r0, v0, r1, v1, add.extrapolation = FALSE)
}
\arguments{
\item{r0}{vector of incremental coordinates in the reference "equivalent" material}

\item{v0}{vector of values in the reference material evaluated at r0}

\item{r1}{vector of incremental coordinates in the actual material}

\item{v1}{vector of values in the actual material evaluated at r1}

\item{add.extrapolation}{evaluate the extrapolation values using a linear fit over EPL(r). If it is FALSE, it will use 1 as angular coefficient (i.e. it will assume that the teo materials are equal outside the specified coordinates: EPL = r)}
}
\value{
a function r0=r0(r1): the epl in the reference material as a function of the length in the actual material.
}
\description{
The evaluation is performed by matching the two cumulative distributions.
}
\seealso{
Other Values Utilities: \code{\link{points2intervals}}
}

