\name{file.variable.gate}
\alias{file.variable.gate}
\title{Map the variable name to file name (Gate)}
\usage{
file.variable.gate(variable)
}
\arguments{
  \item{variable}{The variable name.}
}
\value{
The prefix of the file name.
}
\description{
Map the variable name to file name (Gate)
}
\seealso{
Other PlanGate: \code{\link{create.gate.structure}};
\code{\link{create.plan.gate}};
\code{\link{create.waterbox.gate}};
\code{\link{set.ct.gate}}
}

