% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{get.alpha.beta.beamLUT}
\alias{get.alpha.beta.beamLUT}
\title{Get alpha/beta from beamLUTs (pure-dek)}
\usage{
get.alpha.beta.beamLUT(beamLUTs, beams, x = NULL, y = NULL, z = NULL,
  variables = c("Alpha[Gy^(-1)]", "Beta[Gy^(-2)]"))
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x, y, z}{the coordinates of the 3D array.}

\item{variable}{name vector to override the default name of the variables (c('Alpha[Gy^(-1)]', 'Beta[Gy^(-2)]')).}
}
\value{
A values object (if  x, y and z are specified), otherwise a sparse array.
}
\description{
Get the net alpha and beta Linear-Quadratic parameters distribution (values object or sparse array) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.bdose.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.letd.beamLUT}},
  \code{\link{get.rbe.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{read.beamLUTs}}, \code{\link{read.beamLUT}},
  \code{\link{write.beamLUT}}
}
