\name{get.array.values}
\alias{get.array.values}
\title{estrai array della variabile specificata dai values}
\usage{
get.array.values(values, variable)
}
\description{
estrai array della variabile specificata dai values
}
\seealso{
Other Array Values: \code{\link{add.array.values}};
\code{\link{get.origin}}; \code{\link{set.origin}}
}

