% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{get.bdose.beamLUT}
\alias{get.bdose.beamLUT}
\title{Get equivalent dose from beamLUTs (pure-dek)}
\usage{
get.bdose.beamLUT(beamLUTs, beams, x, y, z,
  variable = "BiologicalDose[Gy(RBE)]", alphaX, betaX)
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x, y, z}{the coordinates of the 3D array.}

\item{variable}{to override the default name of the variable ('Dose[Gy]').}

\item{alphaX, betaX}{the Linear-Quadratic parameters of the reference radiation (they can be two arrays/vectors)}
}
\value{
a values object.
}
\description{
Get the net equivalent (biological) dose distribution (values object) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.letd.beamLUT}},
  \code{\link{get.rbe.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{read.beamLUTs}}, \code{\link{read.beamLUT}},
  \code{\link{write.beamLUT}}
}
