% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.beams}
\alias{get.beams}
\title{Get beams from plan}
\usage{
get.beams(plan, ...)
}
\arguments{
\item{plan}{The plan object.}

\item{input}{If TRUE the input beams are retrieved.}
}
\description{
Get the beams object associated to the plan. There are two type of beams: input and output. Inputs beams indicates the beams used for a forward planning, or for the initial condition of the optimization procedure in the inverse planning. Output beams indicates the beams evaluated in the inverse planning.
}
\seealso{
Other Beams: \code{\link{add.field}};
  \code{\link{convert.fluence2mu}};
  \code{\link{convert.mu2fluence}};
  \code{\link{create.beam}}; \code{\link{create.field}};
  \code{\link{display.beamports}};
  \code{\link{display.beams}};
  \code{\link{get.ct.entrance.point}};
  \code{\link{get.isocenter}}; \code{\link{get.rays}};
  \code{\link{read.beams.fluka}}; \code{\link{read.beams}};
  \code{\link{write.beams}}
}

