% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ct-functions.R
\name{get.ct}
\alias{get.ct}
\title{Get ct object from plan object}
\usage{
get.ct(plan)
}
\arguments{
\item{plan}{The plan object}
}
\description{
Get ct object from plan object
}
\seealso{
Other CT: \code{\link{sanitize.ct}}
}

