% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.dektools.env}
\alias{get.dektools.env}
\title{Get Environment Variables}
\usage{
get.dektools.env()
}
\description{
Get Environment Variables
}
\seealso{
Other Environment: \code{\link{check.gate}};
  \code{\link{check.plankit}};
  \code{\link{set.dektools.env}};
  \code{\link{setenv.rplanit}}
}

