% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.dose.beamLUT}
\alias{get.dose.beamLUT}
\title{Get Dose from beamLUTs (pure-dek)}
\usage{
get.dose.beamLUT(beamLUTs, beams, x, y, z, variable = "Dose[Gy]")
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x,y,z}{the coordinates of the 3D array.}

\item{variable}{to override the default name of the variable ('Dose[Gy]').}
}
\value{
a values object.
}
\description{
Get the net Dose distribution (values object) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}};
  \code{\link{get.bdose.beamLUT}};
  \code{\link{get.letd.beamLUT}};
  \code{\link{get.rbe.beamLUT}};
  \code{\link{get.values.for.beam}};
  \code{\link{read.beamLUT}}; \code{\link{write.beamLUT}}
}

