% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.install.path}
\alias{get.install.path}
\title{Get installing folder}
\usage{
get.install.path()
}
\description{
Get the path to the folder where external precompiled packages (such as plankit, survivals and gate) are/will be installed.
}

