% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{get.letd.beamLUT}
\alias{get.letd.beamLUT}
\title{Get dose averaged LET from beamLUTs (pure-dek)}
\usage{
get.letd.beamLUT(beamLUTs, beams, x = NULL, y = NULL, z = NULL,
  variable = "DoseAveragedLET[keV/um]")
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x, y, z}{the coordinates of the 3D array.}

\item{variable}{to override the default name of the variable ('DoseAveragedLET[keV/um]').}
}
\value{
A values object (if  x, y and z are specified), otherwise a sparse array.
}
\description{
Get the net dose averged LET distribution (values object or sparse array) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}},
  \code{\link{get.bdose.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.rbe.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{read.beamLUT}}, \code{\link{write.beamLUT}}
}

