% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.rbe.beamLUT}
\alias{get.rbe.beamLUT}
\title{Get RBE from beamLUTs (pure-dek)}
\usage{
get.rbe.beamLUT(beamLUTs, beams, x, y, z, variable = "RBE", alphaX, betaX)
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x,y,z}{the coordinates of the 3D array.}

\item{alphaX,betaX}{the Linear-Quadratic parameters of the reference radiation (they can be two arrays/vectors)}

\item{variable}{to override the default name of the variable ('RBE').}
}
\value{
a values object.
}
\description{
Get the net RBE distribution (values object) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}};
  \code{\link{get.bdose.beamLUT}};
  \code{\link{get.beamLUTs}};
  \code{\link{get.dose.beamLUT}};
  \code{\link{get.letd.beamLUT}};
  \code{\link{get.values.for.beam}};
  \code{\link{read.beamLUT}}; \code{\link{write.beamLUT}}
}

