% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.subset.values}
\alias{get.subset.values}
\title{Get a subset of a  \code{values} object.}
\usage{
get.subset.values(values, variables = NULL, xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf), zlim = c(-Inf, Inf))
}
\arguments{
\item{values}{the \code{values} object}

\item{variables}{A vector of the subset of variables to extract from the values object.}

\item{xlim,ylim,zlim}{The ranges of the subset coordinates.}
}
\description{
Get a subset of a  \code{values} object.
}
\seealso{
Other Values Manipulation: \code{\link{add.array.values}};
  \code{\link{get.array.values}}; \code{\link{get.origin}};
  \code{\link{set.origin}}
}

