% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{get.subset.vois}
\alias{get.subset.vois}
\title{Get a subset of a  \code{vois} object.}
\usage{
get.subset.vois(vois, xlim = c(-Inf, Inf), ylim = c(-Inf, Inf),
  zlim = c(-Inf, Inf))
}
\arguments{
\item{xlim, ylim, zlim}{The ranges of the subset coordinates.}

\item{values}{the \code{vois} object}
}
\description{
Get a subset of a  \code{vois} object.
}
\seealso{
Other VOIs: \code{\link{add.vois}},
  \code{\link{create.vois.from.logical.array}},
  \code{\link{create.vois}}, \code{\link{get.voi.logical}},
  \code{\link{get.voiindex}}, \code{\link{get.vois}},
  \code{\link{is.in.voi}}
}
