% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.values.dataframe}
\alias{get.values.dataframe}
\title{recupera matrice di valori in formato dataframe}
\usage{
\method{get.values}{dataframe}(plan)
}
\description{
recupera matrice di valori in formato dataframe
}
\seealso{
Other Values: \code{\link{create.values}};
  \code{\link{get.dataframe.from.values}};
  \code{\link{get.sparse.array.from.values}};
  \code{\link{get.sparse.array.gate}};
  \code{\link{get.values.from.sparse.array}};
  \code{\link{get.values.gate.plan}};
  \code{\link{get.values.plankit.plan}};
  \code{\link{get.values}}; \code{\link{merge.values}};
  \code{\link{sanitize.values}}
}

