\name{get.values.from.sparse.array}
\alias{get.values.from.sparse.array}
\title{ricostruisce oggetto values 3d dalla matrice sparsa}
\usage{
get.values.from.sparse.array(sparse.array, variable = "Dose[Gy]", x, y, z)
}
\description{
l'oggetto values conterrà solo una variabile, con il nome
specificato
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.profile}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.gate}}; \code{\link{get.values}};
\code{\link{merge.values}}; \code{\link{sanitize.values}}
}

