\name{get.values.gate}
\alias{get.values.gate}
\title{Get values object (Gate)}
\usage{
get.values.gate(plan.gate)
}
\arguments{
  \item{plan.gate}{The plan object.}
}
\description{
Get the values matrices for a plan (Gate).
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.sparse.array.gate}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values}}; \code{\link{merge.values}};
\code{\link{sanitize.values}}
}

