% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{get.voiindex}
\alias{get.voiindex}
\title{Get the VOI index from the contours file}
\usage{
get.voiindex(voi, file.contours = NULL, contours = NULL)
}
\arguments{
\item{voi}{the VOI name.}

\item{file.contours}{the contours file.}

\item{contours}{a contours object (data frame).}
}
\value{
the VOI index
}
\description{
Get the VOI index from the contours file
}
\seealso{
Other VOIs: \code{\link{add.vois}},
  \code{\link{create.vois.from.logical.array}},
  \code{\link{create.vois}}, \code{\link{get.subset.vois}},
  \code{\link{get.voi.logical}}, \code{\link{get.vois}},
  \code{\link{is.in.voi}}
}
