\name{get.voxel.index}
\alias{get.voxel.index}
\title{get voxel index from coordinates}
\usage{
get.voxel.index(x, y, z, Xv, Yv, Zv)
}
\arguments{
  \item{x,y,z}{coordinate vectors.}

  \item{Xv,Yv,Zv}{coordinates of the voxels}
}
\value{
A vector.
}
\description{
Merge a list of values objects in a single object. It
assumes that the variable names are unique.
}
\seealso{
Other Values, Utilities: \code{\link{create.intervals}};
\code{\link{generate.values.from.events}};
\code{\link{get.profile}}
}

