% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.zBP.fun}
\alias{get.zBP.fun}
\title{Return a function z.BP(E)}
\usage{
get.zBP.fun(lut.name, EBP = FALSE)
}
\arguments{
\item{lut.name}{The name of the lut.}

\item{EBP}{Return E(z.BP) instead of z.BP(E) if true.}
}
\description{
z.BP is the depth of the Bragg peak as a function of Energy (E). Normally the energy is intended as the specific energy (MeV/u). The function is obtained by interpolating the data contained in the file <lut.name>.zBraggPeaks.1d.
}
\seealso{
Other LUT: \code{\link{lut.add.alpha}};
  \code{\link{lut.add.beta}};
  \code{\link{lut.add.rbe.alpha}};
  \code{\link{lut.radial.integral}};
  \code{\link{read.lut}}; \code{\link{write.lut}}
}

