% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading-writing-functions.R
\name{header2list}
\alias{header2list}
\title{Convert DICOM header plain data.frame to nested list}
\usage{
header2list(header, index = 1, lev = 0)
}
\arguments{
\item{header}{the header data frame.}

\item{index}{the starting row of the header data.frame to be parsed.}

\item{lev}{starting nesting level.}
}
\description{
Convert the DICOM header plain data.frame obtained from the oro.dicom::readDICOMfile() function
into a nested list of data. The nested list is analogous to the structure obtained in MatLab from
the function dicominfo().
}
