% Generated by roxygen2 (4.0.1): do not edit by hand
\name{identify.slice}
\alias{identify.slice}
\title{identifica slice corrispondente alla coordinata}
\usage{
\method{identify}{slice}(x, xvec)
}
\description{
identifica slice corrispondente alla coordinata
}
\seealso{
Other ValuesUtilities: \code{\link{interpolate.values}}
}

