% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install.puredek}
\alias{install.puredek}
\title{Install Pure-dek}
\usage{
install.puredek()
}
\description{
Install Pure-dek
}
\seealso{
Other Install: \code{\link{install.gate}};
  \code{\link{install.plankit}};
  \code{\link{install.survival}}
}

