% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install.survival}
\alias{install.survival}
\title{Install Survival}
\usage{
install.survival()
}
\description{
Install Survival
}
\seealso{
Other Install: \code{\link{install.gate}};
  \code{\link{install.plankit}};
  \code{\link{install.puredek}}
}

