% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{ion.from.z}
\alias{ion.from.z}
\title{Ion from Z}
\usage{
ion.from.z(Z, with.atomic.mass = FALSE)
}
\arguments{
\item{Z}{A numeric vector containing the ion charges.}

\item{with.atomic.mass}{Uses ion names including the atomic mass number (default FALSE.)}
}
\value{
A numeric vector containing the atomic mass numbers.
}
\description{
Returns the name of the ion/atom from its charge (Z).
The name is its element symbol, optionally including
its atomic mass number (e.g.: "C" or "12C" for carbon ion.)
Note: the available range is from 1H to 20 Ne.
}
\seealso{
Other Particles: \code{\link{Zeff}},
  \code{\link{a.from.ion}}, \code{\link{beta.c}},
  \code{\link{remove.atomic.mass}},
  \code{\link{rest.energy}}, \code{\link{z.from.ion}}
}
