% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.in.voi}
\alias{is.in.voi}
\title{Check if point(s) is inside VOI}
\usage{
is.in.voi(x, y, z, contours, voi)
}
\arguments{
\item{x,y,z}{Coordinate(s) of the point.}

\item{contours}{Contours object.}

\item{voi}{The voi name.}
}
\value{
A logical vector.
}
\description{
Check if point(s) is inside VOI
}
\seealso{
Other VOIs: \code{\link{create.vois}};
  \code{\link{get.subset.vois}};
  \code{\link{get.voi.logical}};
  \code{\link{get.voiindex}}; \code{\link{get.vois}}
}

