\name{lut.add.alpha}
\alias{lut.add.alpha}
\title{aggiungi alpha alla lut}
\usage{
lut.add.alpha(lut)
}
\description{
la formula: LethAlphaPerEvent / DosePerEvent
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
\code{\link{lut.add.beta}};
\code{\link{lut.add.rbe.alpha}};
\code{\link{lut.radial.integral}}; \code{\link{read.lut}};
\code{\link{write.lut}}
}

