\name{merge.values}
\alias{merge.values}
\title{Merge a list of values}
\usage{
merge.values(values.list)
}
\arguments{
  \item{values.list}{A list of values objects.}
}
\value{
A values object
}
\description{
Merge a list of values objects in a single object. It
assumes that the variable names are unique.
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.sparse.array.gate}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values.gate}}; \code{\link{get.values}};
\code{\link{sanitize.values}}
}

