% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{mkm.lower}
\alias{mkm.lower}
\title{MKM lower limits}
\format{An object of class \code{numeric} of length 4.}
\usage{
mkm.lower
}
\description{
MKM lower limits
}
\keyword{datasets}
