% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{mkm.upper}
\alias{mkm.upper}
\title{MKM upper limits}
\format{\preformatted{ num [1:4] 1 1 10 1
}}
\usage{
mkm.upper
}
\description{
MKM upper limits
}
\keyword{datasets}

