% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ntcp.kallman}
\alias{ntcp.kallman}
\title{NTCP (Biological Dose, Relative Seriality Model [Kallman1992a])}
\usage{
ntcp.kallman(dvh.dose, D50 = 57, R = 3, gam = 6.7, s = 1, d = 2,
  alphaX = NULL, betaX = NULL, Nf = 30, G = NULL, RBE = 1)
}
\arguments{
\item{dvh.dose}{a dvh object in which is stored the dose (biological dose) for the specific volume.}

\item{D50,R,gam}{tissue specific biological parameters: tolerance dose at 50\%, R=alphaX/betaX, clonogenic density.}

\item{s}{relative seriality parameter of the Kallman's model.}

\item{d}{dose per fraction.}

\item{alphaX,betaX}{alternatively to (D50, R, gam), it is possible to use directly the LQ parameters as biological parameters. These parameters can be vectors.}

\item{Nf}{number of fractions.}

\item{G}{Lea-Catcheside dose-protraction factor (if not define it is assumed complete repair between fractions, G=1/Nf).}

\item{RBE}{Relative Biological Effectivenes.}
}
\value{
The NTCP value
}
\description{
It Uses as default parameters those of the "spinal cord" [from Su2010].
}
\seealso{
Other General TCP/NTCP Models: \code{\link{ntcp.S}};
  \code{\link{ntcp.kallman.train}}; \code{\link{tcp.S}};
  \code{\link{tcp.lqp.train}}; \code{\link{tcp.lqp}}
}

