% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{ntcp.kallman.train}
\alias{ntcp.kallman.train}
\title{NTCP (Biological Dose, Relative Seriality Model [Kallman1992a], Variable fractions)}
\usage{
ntcp.kallman.train(dvhs, D50 = 57, R = 3, gam = 6.7, s = 1, d = 2,
  alphaX = NULL, betaX = NULL, RBE = 1)
}
\arguments{
\item{dvhs}{a list of biological dose DVHs}

\item{D50, R, gam}{tissue specific biological parameters: tolerance dose at 50\%, R=alphaX/betaX, clonogenic density}

\item{s}{relative seriality parameter}

\item{d}{dose per fraction}

\item{RBE}{Relative Biological Effectivenes}

\item{Nf}{number of fractions}
}
\value{
The NTCP value
}
\description{
NTCP evaluation from the biological dose, using arbitrary set of fractions.
}
\details{
The total treatment correspond to a "train" of acute doses distribution (with no time correlation, i.e. full-repair is assumed).
The biological dose distribution, for each voxel, can be different among fractions.
Hence a list of DVH objects is needed. Each element of the list correspond to a DHV evaluated for a specific fraction.

By default it is assumed that the \code{dvh.dose} contains the biological dose dvh [Gy(E)]. Alternatively it could contain the "physical" dose [Gy] and the biological dose is evaluated by specifying the RBE per fraction (assumed to be constant over the irradiated volume).

If alphaX and betaX are not explicity defined, they were deduced from \code{D50}, \code{gam} and \code{R}.
}
\seealso{
Other General TCP/NTCP Models: \code{\link{ntcp.S}},
  \code{\link{ntcp.kallman}}, \code{\link{tcp.S}},
  \code{\link{tcp.lqp.train}}, \code{\link{tcp.lqp}}
}
