% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{NTCP.LL.cMKM}
\alias{NTCP.LL.cMKM}
\title{calcola la log likelihood di complicazioni (NTCP), per un singolo caso, rispetto al dato clinico}
\usage{
NTCP.LL.cMKM(values, vois, alphaX, betaX, rN, rd, beta, particleType, G = G,
  s = s, Np = Np, R = R, Nf = Nf, voi = voi)
}
\description{
Utilizza il modello cMKM.
Assume frazioni indipendenti e costanti.
Assume la presenza di matrice di dose fisica e di LETd per il calcolo LM.
}
\details{
Np -> numero di pazienti
R -> numero di pazienti responder (R=Np*NTCP)
Nf -> numero di frazioni
}
\seealso{
Other Likelihood Evaluations: \code{\link{NTCP.LL.LM}};
  \code{\link{TCP.LL.LM}}; \code{\link{TCP.LL.cMKM}}
}

