% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{NTCP.LM}
\alias{NTCP.LM}
\title{calcola NTCP con modello LM}
\usage{
NTCP.LM(values, vois, voi = "PTV", alpha0, beta, m, G = G, s = s,
  Nf = Nf)
}
\description{
calcola NTCP con modello LM
}
\seealso{
Other LM/cMKM TCP/NTCP Models: \code{\link{NTCP.cMKM}},
  \code{\link{TCP.LM}}, \code{\link{TCP.cMKM}}
}

