% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{ntcp.S}
\alias{ntcp.S}
\title{NTCP (Relative Seriality Model [Kallman1992a])}
\usage{
ntcp.S(dvh.survival, gam = 2, s = 1, Nf = 30)
}
\arguments{
\item{dvh.survival}{DVH of the cell "survivals" for a single fraction}

\item{gam}{radiobiological parameter, related to the clonogenic density}

\item{s}{relative serialiy parameter}

\item{Nf}{number of fractions}
}
\value{
The NTCP value.
}
\description{
Evaluate the NTCP directly from the DVH of the "survivals" (for a given endpoint and VOI), without using the biological dose and/or indirect radiobiological parameters for a reference radiation.
}
\details{
It assumes identical and uncorrelated fractions (complete-repair)
}
\seealso{
Other General TCP/NTCP Models: \code{\link{ntcp.kallman.train}},
  \code{\link{ntcp.kallman}}, \code{\link{tcp.S}},
  \code{\link{tcp.lqp.train}}, \code{\link{tcp.lqp}}
}
\concept{General TCP/NTCP Models}
