\name{read.3d.spliced}
\alias{read.3d.spliced}
\title{Legge file in formato "3d" e restituisce un dataframe}
\usage{
read.3d.spliced(file.name, x.min = -Inf, x.max = +Inf, y.min = -Inf,
  y.max = +Inf, z.min = -Inf, z.max = +Inf, Nsplices = 16)
}
\description{
La trasformazione in dataframe viene fatta a "splice" per
risparmiare memoria
}
\details{
OBSOLETO
}
\seealso{
Other R/W Arrays: \code{\link{read.1d}};
\code{\link{read.3d.df}}; \code{\link{read.3d.dicom}};
\code{\link{read.3d.hdr}}; \code{\link{read.3d}}
}

