% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.beamLUT}
\alias{read.beamLUT}
\title{Read the beamLUTs (pure-dek)}
\usage{
read.beamLUT(beamLUT.name, Nx, Ny, Nz)
}
\arguments{
\item{beamLUT.name}{the prefix of the beamLUT files. The complete filenames are expected to be: <beamLUT.name>_<beamIndex>.beamLUT}
}
\description{
Read the beamLUTs (pure-dek)
}

