\name{read.beams.fluka}
\alias{read.beams.fluka}
\title{Read beams from file (Fluka)}
\usage{
read.beams.fluka(fluka.file)
}
\arguments{
  \item{fluka.file}{The file name.}
}
\value{
a beams object.
}
\description{
Read RtIonPlan file (Fluka format) and return a beams
object.
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.beam}}; \code{\link{create.field}};
\code{\link{display.beamports}};
\code{\link{display.beams}}; \code{\link{get.beams}};
\code{\link{get.ct.entrance.point}};
\code{\link{get.isocenter}}; \code{\link{get.rays}};
\code{\link{read.beams}}; \code{\link{write.beams}}
}

