\name{read.beams}
\alias{read.beams}
\title{legge dati e crea oggetto beams}
\usage{
read.beams(beams.file)
}
\description{
legge dati e crea oggetto beams
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.beam}}; \code{\link{create.field}};
\code{\link{display.beams}}; \code{\link{get.beams}};
\code{\link{get.isocenter}}; \code{\link{write.beams}}
}

