% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{read.contours.dicom}
\alias{read.contours.dicom}
\title{Read contours from DICOM file}
\usage{
read.contours.dicom(file.contours.dicom, folder.CT.dicom = NULL,
  CT = NULL, z.CT = NULL, tissue = "R3PIDEd0.2v0+MKM")
}
\arguments{
\item{file.contours.dicom}{the name of the contours file}

\item{folder.CT.dicom}{the name of the DICOM CT folder. The CT is needed to properly identify the slice for each contour polygon.}

\item{CT}{the CT object (alternative to folder.CT.dicom and z.CT)}

\item{z.CT}{a vector for the z coordinates of the slices of the reference CT (alternative to folder.CT.dicom and CT)}
}
\value{
A \code{contours} dataframe consisting of:
\item{id}{index of the contours}
\item{polygon}{index of the polygon}
\item{slice}{index of the slice}
\item{x,y,z}{x,y,z coordinates}
\item{contour}{contour (VOI) name}
\item{tissue}{tissue+biological model name}
\item{type}{contour type (target, OAR, etc.)}
\item{colour}{the hex specification of the colour associated to the contour.}
}
\description{
Read the contours from a DICOM file and return a \code{contours} data frame. The CT data is also mandatory
to identify the slice for each contour polygon. The CT data can be specified with one of the following:
folder.CT.dicom, CT or z.CT. z.CT has precendence to CT that has precendence to folder.CT.dicom.
}
\seealso{
Other R/W Contours: \code{\link{read.contours}},
  \code{\link{write.contours}}
}
\concept{R/W Contours}
