% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{read.contours}
\alias{read.contours}
\title{Read contours from file}
\usage{
read.contours(file.contours, file.CT)
}
\arguments{
\item{file.contours}{the name of the contours file}

\item{file.CT}{the name of the CT file. The CT file (in PlanKIT format) is needed to properly reconstruc the 3D coordinates of the contours.}
}
\value{
A \code{contours} dataframe consisting of:
\item{id}{index of the contours}
\item{polygon}{index of the polygon}
\item{slice}{index of the slice}
\item{x,y,z}{x,y,z coordinates}
\item{contour}{contour (VOI) name}
\item{tissue}{tissue+biological model name}
\item{type}{contour type (target, OAR, etc.)}
}
\description{
Read the contours from a "*.contours" file format (PlanKIT format) and return a \code{contours} dataframe.
}
\seealso{
Other R/W Contours: \code{\link{write.contours}}
}

