% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{read.vois}
\alias{read.vois}
\title{Legge i vois da un file .3d e restituisce un dataframe "esteso"}
\usage{
read.vois(file.name, x.min = -Inf, x.max = +Inf, y.min = -Inf,
  y.max = +Inf, z.min = -Inf, z.max = +Inf)
}
\description{
OBSOLETE
}
\seealso{
Other R/W VOIs: \code{\link{read.vois.array}};
  \code{\link{read.vois.spliced}}
}

