% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{rest.energy}
\alias{rest.energy}
\title{Rest energy}
\usage{
rest.energy(A)
}
\arguments{
\item{A}{A numeric vector containing the atomic mass number (or mass in atomic mass units).}
}
\value{
the rest energies.
}
\description{
Returns the (approximate) rest energy from the atomic mass number.
}
\seealso{
Other Particles: \code{\link{Zeff}},
  \code{\link{a.from.ion}}, \code{\link{beta.c}},
  \code{\link{ion.from.z}},
  \code{\link{remove.atomic.mass}},
  \code{\link{z.from.ion}}
}
\concept{Particles}
