\name{Rplanit-package}
\alias{Rplanit-package}
\alias{Rplanit}
\docType{package}
\title{
RT Ion planning simulations and analysis
}
\description{
A collection of methods and procedures to perform simulations and
    analysis related to treatment planning (with ion beams). It includes a set
    of radiobiological evaluations and display possibilities.
}
\details{
\tabular{ll}{
Package: \tab Rplanit\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-26\cr
License: \tab GPL2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Andrea Attili

Maintainer: Andrea Attili <attili@to.infn.it>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
simple examples of the most important functions
}
