% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plankit-functions.R
\name{run.dek.forward}
\alias{run.dek.forward}
\title{Forward planning (Pure-dek)}
\usage{
run.dek.forward(plan, outmessages = FALSE)
}
\arguments{
\item{plan}{the Plan object. It should include also the specification of the beams (i.e. the "RTIonPlan").}

\item{outmessages}{if TRUE, it displays the messages from the Pure-dek output.}
}
\value{
a Plan Object (including data/information of the evaluated output of the forward planning, e.g. values, vois, etc.)
}
\description{
Perform a forward planning with Pure-dek. It assumes that the beams are defined in theplan object, checking in this order: beams, inoutBeamsFile, outputBeamsFile.
}
\seealso{
Other PlanKIT: \code{\link{embed.data}};
  \code{\link{run.dek.inverse}}
}

