% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankit-functions.R
\name{run.dek.inverse}
\alias{run.dek.inverse}
\title{Inverse planning (Pure-dek)}
\usage{
run.dek.inverse(plan, outmessages = FALSE)
}
\arguments{
\item{plan}{the Plan object.}

\item{outmessages}{if TRUE, it displays the messages from Pure-dek output.}
}
\value{
a Plan Object (including data/information of the evaluated output of the inverse planning, e.g. beams, values, vois, etc.)
}
\description{
Perform an inverse planning with Pure-dek.
}
\seealso{
Other PlanKIT: \code{\link{embed.data}},
  \code{\link{run.dek.forward}}
}

