% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vois-functions.R
\name{sanitize.contours}
\alias{sanitize.contours}
\title{Correct consistency of contours}
\usage{
sanitize.contours(contours)
}
\arguments{
\item{contours}{the contours dataframe}
}
\description{
Check and correct the consistency of a contours object (such as not uniform and/or ordered polygon numbering, negative index, etc.)
}
\seealso{
Other Contours: \code{\link{add.colours.contours}},
  \code{\link{display.contours.legend}},
  \code{\link{get.contours}}, \code{\link{volume.contours}}
}
