% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-functions.R
\name{sanitize.ct}
\alias{sanitize.ct}
\title{rimuove valori fuori range della ct}
\usage{
sanitize.ct(ct, HU.min = -1000, HU.max = 3000, threshold.air = NULL)
}
\description{
rimuove valori fuori range della ct
}
\seealso{
Other CT: \code{\link{get.ct}}
}
