\name{sanitize.values}
\alias{sanitize.values}
\title{rimuove valori nulli non "fisici"}
\usage{
sanitize.values(values)
}
\description{
rimuove valori nulli non "fisici"
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.profile}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values}}
}

