% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{set.lut.folder}
\alias{set.lut.folder}
\title{Set LUT folder}
\usage{
set.lut.folder(lut.folder = NULL)
}
\arguments{
\item{lut.folder}{The lut.folder. If it is not defined the lut folder is set to the default one (the pure-dek local installation.)}
}
\description{
Set the LUT (look up table) folder to be used in pure-dek computations.
}
\seealso{
Other Environment: \code{\link{check.gate}},
  \code{\link{check.plankit}},
  \code{\link{get.dektools.env}},
  \code{\link{set.dektools.env}},
  \code{\link{set.wepl.folder}},
  \code{\link{setenv.rplanit}}
}
\concept{Environment}
