% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{set.wepl.folder}
\alias{set.wepl.folder}
\title{Set WEPL folder}
\usage{
set.wepl.folder(wepl.folder = NULL)
}
\arguments{
\item{wepl.folder}{The wepl.folder. If it is not defined the lut folder is set to the default one (the pure-dek local installation.)}
}
\description{
Set the WEPL (water equivalent path length) folder where the wepl files are stored.
}
\seealso{
Other Environment: \code{\link{check.gate}},
  \code{\link{check.plankit}},
  \code{\link{get.dektools.env}},
  \code{\link{set.dektools.env}},
  \code{\link{set.lut.folder}},
  \code{\link{setenv.rplanit}}
}
