% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{sparse.array.from.values}
\alias{sparse.array.from.values}
\title{recupera matrice sparsa da values}
\usage{
sparse.array.from.values(values, variable = "Dose[Gy]", threshold = 0)
}
\description{
accetta values con più di una variabile. la selezione è fatta sulla prima.
(che può coincidere o meno con la variabile specificata)
}
\seealso{
Other Values: \code{\link{create.values}};
  \code{\link{dataframe.from.values}};
  \code{\link{get.sparse.array.gate}};
  \code{\link{get.values.dataframe}};
  \code{\link{get.values.gate.plan}};
  \code{\link{get.values.plankit.plan}};
  \code{\link{get.values}}; \code{\link{merge.values}};
  \code{\link{sanitize.values}};
  \code{\link{values.from.sparse.array}}
}

