% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{srim.let.min}
\alias{srim.let.min}
\title{srim.limits}
\format{\preformatted{ Named num [1:10] 0.206 0.824 1.9 3.351 5.484 ...
 - attr(*, "names")= chr [1:10] "H" "He" "Li" "Be" ...
}}
\usage{
srim.let.min
}
\description{
srim.limits
}
\keyword{datasets}

