% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{TCP.cMKM}
\alias{TCP.cMKM}
\title{calcola TCP con modello cMKM}
\usage{
TCP.cMKM(values = values, vois = vois, voi = "PTV", alphaX = alphaX,
  betaX = betaX, rN = rN, rd = rd, beta = beta,
  particleType = particleType, G = G, Nf = Nf)
}
\description{
calcola TCP con modello cMKM
}
\seealso{
Other LM/cMKM TCP/NTCP Models: \code{\link{NTCP.LM}};
  \code{\link{NTCP.cMKM}}; \code{\link{TCP.LM}}
}

