% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{tcp.lqp}
\alias{tcp.lqp}
\title{TCP (Poissonian Model, Biological Dose)}
\usage{
tcp.lqp(dvh.dose = dvh.dose, D50 = 66, R = 10, gam = 2, d = 2,
  alphaX = NULL, betaX = NULL, Nf = 30, G = NULL, RBE = 1)
}
\arguments{
\item{dvh.dose}{a dvh object in which is stored the dose (biological dose) for the specific volume.}

\item{D50,R,gam}{tissue specific biological parameters: tolerance dose at 50\%, R=alphaX/betaX, clonogenic density.}

\item{d}{dose per fraction.}

\item{alphaX,betaX}{alternatively to (D50, R, gam), it is possible to use directly the LQ parameters as biological parameters. These parameters can be vectors.}

\item{Nf}{number of fractions.}

\item{G}{Lea-Catcheside dose-protraction factor (if not define it is assumed complete repair between fractions, G=1/Nf).}

\item{RBE}{Relative Biological Effectivenes.}
}
\value{
The TCP value
}
\description{
Evaluate the Tumor Control Probability (TCP) using the Poissonian Model from the knowledge of the biological dose distribution.
}
\details{
The default parameters are taken from the example in [in Bentzen1997]
plus the R=alphaX/betaX parameter for tumoral tissues.

By default it is assumed that the \code{dvh.dose} contains the biological dose dvh [Gy(E)]. Alternatively it could contain the "physical" dose [Gy] and the biological dose is evaluated by specifying the RBE per fraction (assumed to be constant over the irradiated volume).

If alphaX and betaX are not explicity defined, they were deduced from \code{D50}, \code{gam} and \code{R}
}
\seealso{
Other General TCP/NTCP Models: \code{\link{ntcp.S}};
  \code{\link{ntcp.kallman.train}};
  \code{\link{ntcp.kallman}}; \code{\link{tcp.S}};
  \code{\link{tcp.lqp.train}}
}

