% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{threshold.values}
\alias{threshold.values}
\title{Set values under threshold to NA}
\usage{
threshold.values(values, threshold = 0, variable = "Dose[Gy]")
}
\arguments{
\item{values}{a values object}

\item{threshold}{the threshold value}

\item{the}{variable associated to the threshold}
}
\value{
a values object with NA where the specified variable is equal or less the threshold value
}
\description{
The treshold is definied for a single variable. Where the specified quantities is equal or less the threshold, all variables are set to NA.
}
\seealso{
Other Values: \code{\link{combine.values}},
  \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{merge.values}},
  \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{values.from.sparse.array}}
}
\concept{Values}
