% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{values.add.RBE.alpha}
\alias{values.add.RBE.alpha}
\title{Calcola l'RBE.alpha e l'aggiunge in values}
\usage{
values.add.RBE.alpha(values, model = NULL, alphaX = 0.2,
  model.LM = NULL, model.cMKM = NULL)
}
\description{
viene passato come argomento il modello:
(NULL -> usa "Alpha[Gy^(-1)]")
(LM -> usa "Alpha.LM[Gy^(-1)]")
(cMKM -> usa "Alpha.cMKM[Gy(-1)]")
}
\details{
assume che la variabile alpha sia già presente in values, atrimenti la calcola
esplicitamente.

nota: nell'oggetto values NON deve essere già presente la variabile "Survival"

nota: nei parametri del modello deve essere esplcitata obbligatoriamente anche il beta
anche se non viene esplicitamente usatato, viene usato per calcolare le sopravvivenze
}
\seealso{
Other Values Bioogical Evaluations: \code{\link{values.add.RBE}},
  \code{\link{values.add.survival.LM}},
  \code{\link{values.add.survival.cMKM}}
}
\concept{Values Bioogical Evaluations}
