% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\name{values.add.survival.cMKM}
\alias{values.add.survival.cMKM}
\title{Calcola la sopravvivenza secondo il modello cMKM e la aggiunge in values}
\usage{
values.add.survival.cMKM(values = values, alphaX = alphaX,
  betaX = betaX, rN = rN, rd = rd, particleType = particleType,
  beta = beta, add.alpha = TRUE)
}
\description{
alphaX, betaX, rN, rd sono i parametri del modello MKM
particleType -> ione di riferimento
beta -> valore di beta (indipendente dal LETd)
Nota: l'oggetto values deve avere le due variabili 'Dose[Gy]' e 'DoseAveragedLET[keV/um]'
}
\seealso{
Other Values Bioogical Evaluations: \code{\link{values.add.RBE.alpha}},
  \code{\link{values.add.RBE}},
  \code{\link{values.add.survival.LM}}
}
\concept{Values Bioogical Evaluations}
