% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{values.from.sparse.array}
\alias{values.from.sparse.array}
\title{Create values from sparse array}
\usage{
values.from.sparse.array(sparse.array, variables = NULL, x, y, z)
}
\arguments{
\item{sparse.array}{The sparse array. Note: the sparse array is a dataframe containing at least the following columns: 'voxel.index' and 'value'. Optionally it could contain also the column 'variable'. The latter is mandatory if there are multiple variables.}

\item{variables}{The name of the variable(s). If there is only one variable specified and the sparse array dose not contain the column "variable", it is assumed that all the value correspond to the specified variable. It it is NULL, the values object will contains all the variable found in the sparse array.}
}
\value{
Values object.
Note: the sparse array is a dataframe containing at least the following columns: 'voxel.index' and 'value'
}
\description{
Create a values object from a sparse array.
}
\seealso{
Other Values: \code{\link{combine.values}},
  \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}}
}

