\name{volume.contours}
\alias{volume.contours}
\title{Evaluate volume from contours}
\usage{
volume.contours(contours, voi = NULL)
}
\arguments{
  \item{contours}{the contours dataframe}
}
\description{
Evaluate the volume defined by its contours.
}
\seealso{
Other Contours: \code{\link{get.contours}}
}

