% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.3d.array}
\alias{write.3d.array}
\title{Write 3D array (Pure-dek)}
\usage{
write.3d.array(values, file.name)
}
\arguments{
\item{values}{the \code{values} object}

\item{file.name}{the file name}
}
\description{
Write a \code{values} object in a file using the pure-dek format (*.3d)
}

