% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading-writing-functions.R
\name{write.3d}
\alias{write.3d}
\title{Write 3D array (Pure-dek)}
\usage{
write.3d(values, file.name)
}
\arguments{
\item{values}{the \code{values} object}

\item{file.name}{the file name}
}
\description{
Write a \code{values} object in a file using the pure-dek format (*.3d)
}

