\name{write.beams}
\alias{write.beams}
\title{salva tabella beams (RTIonPlan)}
\usage{
write.beams(beams, file.name, format = "plankit", ion = "1H")
}
\description{
Puo salvare in diversi formati (default PlanKIT)
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.beam}}; \code{\link{create.field}};
\code{\link{display.beams}}; \code{\link{get.beams}};
\code{\link{get.isocenter}}; \code{\link{read.beams}}
}

