% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{write.lut.4d}
\alias{write.lut.4d}
\title{Write LUT 4d data array}
\usage{
write.lut.4d(lut.name, lut)
}
\arguments{
\item{lut.name}{The name (file prefix) of the LUT}

\item{lut}{the LUT object containing the 4d data array}
}
\description{
Write LUT 4d data array
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}},
  \code{\link{lut.add.alpha}}, \code{\link{lut.add.beta}},
  \code{\link{lut.add.rbe.alpha}},
  \code{\link{lut.radial.integral}},
  \code{\link{read.lut}}, \code{\link{write.lut.meta}},
  \code{\link{write.lut.optics}},
  \code{\link{write.lut.phasespace}},
  \code{\link{write.lut.zBP}}, \code{\link{write.lut}}
}
\concept{LUT}
