% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{write.lut.meta}
\alias{write.lut.meta}
\title{Write LUT meta file}
\usage{
write.lut.meta(lut.name, physical.variables = "DosePerEvent[Gy/primary]",
  biological.variables = c("LethAlphaPerEvent[1/primary]",
  "SqrtLethBetaPerEvent[1/primary]"), tissue.names = "NULL", alphaX = 1,
  betaX = 1)
}
\arguments{
\item{lut.name}{The name (file prefix) of the LUT}

\item{physical.variables}{Names of the physical variables}

\item{biological.variables}{Names of the biological variables}

\item{alphaX,betaX}{vectors of the LQ parameters for the reference radiation (used to compute the RBE.)}

\item{model.names}{Names of the biological models}
}
\description{
Write LUT meta file
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
  \code{\link{lut.add.alpha}}; \code{\link{lut.add.beta}};
  \code{\link{lut.add.rbe.alpha}};
  \code{\link{lut.radial.integral}};
  \code{\link{read.lut}}; \code{\link{write.lut.4d}};
  \code{\link{write.lut.optics}};
  \code{\link{write.lut.phasespace}};
  \code{\link{write.lut.zBP}}; \code{\link{write.lut}}
}

