% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{write.lut.optics}
\alias{write.lut.optics}
\title{Write LUT optics (meta) file}
\usage{
write.lut.optics(lut.name, x.magnet.distance = 6000,
  y.magnet.distance = 6000, particle = "1H")
}
\arguments{
\item{lut.name}{The name (file prefix) of the LUT}

\item{x.magnet.distance,y.magnet.distance}{Distances of the scanning magnets.}

\item{particle}{The primary particle.}
}
\description{
Write LUT optics (meta) file
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
  \code{\link{lut.add.alpha}}; \code{\link{lut.add.beta}};
  \code{\link{lut.add.rbe.alpha}};
  \code{\link{lut.radial.integral}};
  \code{\link{read.lut}}; \code{\link{write.lut.4d}};
  \code{\link{write.lut.meta}};
  \code{\link{write.lut.phasespace}};
  \code{\link{write.lut.zBP}}; \code{\link{write.lut}}
}

