% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{write.lut.phasespace}
\alias{write.lut.phasespace}
\title{Write LUT phase-space file}
\usage{
write.lut.phasespace(lut.name, nominalEnergy = NULL, meanEnergy = NULL,
  sigmaEnergy = 0, sigmaX = 4, sigmaY = 4, sigmaX1 = 0, sigmaY1 = 0,
  rhoXX1 = 0, rhoYY1 = 0)
}
\arguments{
\item{lut.name}{The name (file prefix) of the LUT}

\item{lut}{the LUT object containing the 4d data array}
}
\description{
Note: this implementations assumes a "complete" pencilbeam approach, without sub-pencilbeams. Hence the phase space is essentially
a place holder.
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
  \code{\link{lut.add.alpha}}; \code{\link{lut.add.beta}};
  \code{\link{lut.add.rbe.alpha}};
  \code{\link{lut.radial.integral}};
  \code{\link{read.lut}}; \code{\link{write.lut.4d}};
  \code{\link{write.lut.meta}};
  \code{\link{write.lut.optics}};
  \code{\link{write.lut.zBP}}; \code{\link{write.lut}}
}

