% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{write.lut}
\alias{write.lut}
\title{salva la lut a partire dai dati in memoria}
\usage{
write.lut(lut.array, variables = NULL, E = NULL, x = NULL, y = NULL,
  zn = NULL, filename = "lut.4d", binary = FALSE)
}
\description{
salva la lut a partire dai dati in memoria
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}},
  \code{\link{lut.add.alpha}}, \code{\link{lut.add.beta}},
  \code{\link{lut.add.rbe.alpha}},
  \code{\link{lut.radial.integral}},
  \code{\link{read.lut}}, \code{\link{write.lut.4d}},
  \code{\link{write.lut.meta}},
  \code{\link{write.lut.optics}},
  \code{\link{write.lut.phasespace}},
  \code{\link{write.lut.zBP}}
}

