% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{Zeff}
\alias{Zeff}
\title{Evaluate the effective charge}
\usage{
Zeff(Ec, Z, A)
}
\arguments{
\item{Ec}{A numeric vector containing kinetic energies (MeV).}

\item{Z}{A numeric vector containing the charges of the particle.}

\item{A}{A numeric vector containing the atomic mass numbers.}
}
\value{
The effective charge.
}
\description{
Evaluate the effective charge (Zeff) using the Barkas approximate formula.
}
\seealso{
Other Particles: \code{\link{a.from.ion}},
  \code{\link{beta.c}}, \code{\link{ion.from.z}},
  \code{\link{remove.atomic.mass}},
  \code{\link{rest.energy}}, \code{\link{z.from.ion}}
}
