% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.array.values}
\alias{add.array.values}
\title{Add an array to values}
\usage{
add.array.values(values, new.array, variable = "New Data")
}
\arguments{
\item{values}{the \code{values} object}

\item{new.array}{the array to be added}

\item{variable}{the name of the new variable}
}
\description{
Add an array of a new variable to a \code{values} object.
}
\seealso{
Other Values Manipulation: \code{\link{get.array.values}};
  \code{\link{get.origin}};
  \code{\link{get.subset.values}}; \code{\link{set.origin}}
}

