% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.array.values}
\alias{add.array.values}
\title{aggiungi array di una nuova variabile all'oggetto values}
\usage{
add.array.values(values, new.array, variable = "New Data")
}
\description{
aggiungi array di una nuova variabile all'oggetto values
}
\seealso{
Other Values Manipulation: \code{\link{get.array.values}};
  \code{\link{get.origin}}; \code{\link{set.origin}}
}

