% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.field}
\alias{add.field}
\title{Add field ID column to beams}
\usage{
add.field(beams, numeric = TRUE)
}
\arguments{
\item{beams}{the beams data.frame}

\item{numeric}{use a numeric ID for the field}
}
\description{
An unique field ID is associated to a unique combination of gantry and patient angle.
}
\seealso{
Other Beams: \code{\link{convert.fluence2mu}};
  \code{\link{convert.mu2fluence}};
  \code{\link{create.beam}}; \code{\link{create.field}};
  \code{\link{display.beamports}};
  \code{\link{display.beams}}; \code{\link{get.beams}};
  \code{\link{get.ct.entrance.point}};
  \code{\link{get.isocenter}}; \code{\link{get.rays}};
  \code{\link{read.beams.fluka}}; \code{\link{read.beams}};
  \code{\link{write.beams}}
}

