% Generated by roxygen2 (4.0.2): do not edit by hand
\name{available.variables.gate}
\alias{available.variables.gate}
\title{Available variables (Gate)}
\usage{
available.variables.gate()
}
\value{
A list of the available variables
}
\description{
The available variables name identifiers that can be evaluated with Gate.
}

