% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankit-functions.R
\name{available.variables}
\alias{available.variables}
\title{Available variables (Pure-dek)}
\usage{
available.variables()
}
\value{
A list of the available variables
}
\description{
The available variables name identifiers that can be evaluated with PlanKIT
}
\seealso{
Other PlanKITVariables: \code{\link{available.beamlines}},
  \code{\link{available.constraint.types}}
}
\concept{PlanKITVariables}
