% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.gate}
\alias{check.gate}
\title{Check the configuration of Gate}
\usage{
check.gate(use.warning = FALSE)
}
\arguments{
\item{use.warning}{if TRUE it uses warning instead of an error.}
}
\description{
If the check fails it generates an error.
#'
}
\seealso{
Other Environment: \code{\link{check.plankit}};
  \code{\link{get.dektools.env}};
  \code{\link{set.dektools.env}};
  \code{\link{setenv.rplanit}}
}

