% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{check.plankit}
\alias{check.plankit}
\title{Check the configuration of PlanKIT}
\usage{
check.plankit(use.warning = FALSE)
}
\arguments{
\item{use.warning}{if TRUE it uses warning instead of an error.}
}
\description{
If the check fails it generates an error.
}
\seealso{
Other Environment: \code{\link{check.gate}};
  \code{\link{get.dektools.env}};
  \code{\link{set.dektools.env}};
  \code{\link{set.lut.folder}};
  \code{\link{setenv.rplanit}}
}

