% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{combine.values}
\alias{combine.values}
\title{Merge a list of values}
\usage{
combine.values(values.list)
}
\arguments{
\item{values.list}{A list of values objects.}
}
\value{
A values object
}
\description{
Merge a list of values objects in a single object. It assumes that the variable names are not overlapping among different values objects.
}
\seealso{
Other Values: \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{values.from.sparse.array}}
}

