\name{convert.fluence2mu}
\alias{convert.fluence2mu}
\title{converte fluenze in beams da numero di particelle a MU}
\usage{
convert.fluence2mu(beams, type = "protonCnao")
}
\description{
secondo le curve di calibrazione specificate
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.mu2fluence}};
\code{\link{create.beam}}; \code{\link{create.field}};
\code{\link{display.beamports}};
\code{\link{display.beams}}; \code{\link{get.beams}};
\code{\link{get.isocenter}};
\code{\link{read.beams.fluka}}; \code{\link{read.beams}};
\code{\link{write.beams}}
}

