\name{create.beam}
\alias{create.beam}
\title{crea tabella vuota x 1 beam}
\usage{
create.beam()
}
\description{
crea tabella vuota x 1 beam
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.field}}; \code{\link{display.beams}};
\code{\link{get.beams}}; \code{\link{get.isocenter}};
\code{\link{read.beams}}; \code{\link{write.beams}}
}

