% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.field}
\alias{create.field}
\title{Create a field object (dataframe)}
\usage{
create.field(N = 1, beamLine = "protonSimple", targetVOI = "PTV",
  targetVOIIndex = NA, iecGantryAngle = 0, iecPatientSupportAngle = 0,
  interSpotSpacing = c(1.5, 1.5, 1), spotsExtensionOutsideTarget = 0,
  targetIsocenter = c(NA, NA, NA))
}
\arguments{
\item{N}{number of fields.}

\item{targetVOIIndex}{the index of the target VOI.}

\item{iecGantryAngle}{gantry angle.}

\item{iecPatientSupportAngle}{patient support angle.}

\item{interSpotSpacing}{spacing among spots along the three axis of the beam port (mm).}

\item{spotsExtensionOutsideTarget}{thickness of the extension of the spot coverage outside the target VOI.}

\item{targetIsocenter}{coordinates of the isocenter of the target VOI.}
}
\description{
The field object is a dataframe. Different fields are added by apending rows to the dataframe. Each of the following parameter is a dataframe column.
}
\seealso{
Other Beams: \code{\link{add.field}};
  \code{\link{convert.fluence2mu}};
  \code{\link{convert.mu2fluence}};
  \code{\link{create.beam}};
  \code{\link{display.beamports}};
  \code{\link{display.beams}}; \code{\link{get.beams}};
  \code{\link{get.ct.entrance.point}};
  \code{\link{get.isocenter}}; \code{\link{get.rays}};
  \code{\link{read.beams.fluka}}; \code{\link{read.beams}};
  \code{\link{write.beams}}
}

