% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gate-functions.R
\name{create.gate.structure}
\alias{create.gate.structure}
\title{Create Gate structure}
\usage{
create.gate.structure(plan)
}
\arguments{
\item{plan}{A Plan Object.}
}
\description{
Create the folder structure of a Gate simulation for a specific plan.
}
\seealso{
Other PlanGate: \code{\link{create.plan.gate}};
  \code{\link{create.waterbox.gate}};
  \code{\link{file.variable.gate}};
  \code{\link{run.gate.forward}}; \code{\link{set.ct.gate}}
}

