\name{create.intervals}
\alias{create.intervals}
\title{Generate bin intervals from points}
\usage{
create.intervals(x)
}
\arguments{
  \item{x}{A vector.}
}
\value{
A vector (with N+1 elements)
}
\description{
Merge a list of values objects in a single object. It
assumes that the variable names are unique.
}
\seealso{
Other Values, Utilities:
\code{\link{generate.values.from.events}};
\code{\link{get.profile}}; \code{\link{get.voxel.index}}
}

