% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.plan.gate}
\alias{create.plan.gate}
\title{Generate Plan template (Gate)}
\usage{
create.plan.gate(name = "gate.simulation", ctFile = NULL, ctWater = NULL,
  ct = NULL, materialDatabase = "./data/MyMaterials.db",
  HUToMaterialFile = "data/ct-HU2mat.txt", contours = NULL, origin = c(0,
  0, 0), waterIonisationPotential = 78,
  computingValues = c("DoseToMaterial[Gy]"), saveEveryNSeconds = 200,
  size = "auto", position = c(0, 0, 0), computingGridVoxelSizes = c(2, 2,
  2), beams = NULL, beamsFile.gate = "./data/beams.gate",
  particleType = "proton",
  sourceDescriptionFile = "./data/ProtonSimple.txt",
  totalNumberOfPrimaries = 0, K = NULL)
}
\description{
Generate a default (empty) template of the Plan, to be used with Gate.
(documentation for the parameters coming soon...)
}
\seealso{
Other PlanGate: \code{\link{create.gate.structure}};
  \code{\link{create.waterbox.gate}};
  \code{\link{file.variable.gate}};
  \code{\link{run.gate.forward}}; \code{\link{set.ct.gate}}
}

