% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.R
\name{display.beamports}
\alias{display.beamports}
\title{Beam-port splot}
\usage{
display.beamports(beams, plan = NULL, numeric = FALSE, show.plot = TRUE,
  file.name = NULL, height = 7, width = 7, dpi = 300,
  field.names = NULL)
}
\arguments{
\item{beams}{beams dataframe.}

\item{plan}{The plan object.}

\item{numeric}{Display a numeric ID for the fields (default uses a beamLine+angles code).}

\item{file.name}{File name for saving the plot.}

\item{height}{The height of the saved plot image (inches).}

\item{width}{The width of the saved plot image (inches).}

\item{dpi}{The dpi of the saved image.}

\item{field.names}{Optional vector of the names for the different fields}

\item{shot.plot}{Show plot.}
}
\value{
If show.plot=FALSE, it returns a ggplot object.
}
\description{
Beam-port splot
}
\seealso{
Other Beams: \code{\link{add.field}},
  \code{\link{convert.fluence2mu}},
  \code{\link{convert.mu2fluence}},
  \code{\link{create.beam}}, \code{\link{create.field}},
  \code{\link{display.beams}}, \code{\link{get.beams}},
  \code{\link{get.ct.entrance.point}},
  \code{\link{get.isocenter}}, \code{\link{get.rays}},
  \code{\link{read.beams.fluka}}, \code{\link{read.beams}},
  \code{\link{write.beams}}
}
