% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.R
\name{display.profile}
\alias{display.profile}
\title{Display profile}
\usage{
display.profile(profile.values, profile.ct = NULL, profile.names = NULL,
  depth.lim = NULL, show.plot = TRUE, file.name = NULL, height = 7,
  width = 7, return.dataframe = FALSE)
}
\description{
Deprecated: see "display.profiles"
}
\seealso{
Other Profiles: \code{\link{display.profiles}}
}
