\name{display.slice.ct}
\alias{display.slice.ct}
\title{display della ct (usa la funzione display.slice)}
\usage{
display.slice.ct(ct, roi = NULL, x = NA, y = NA, z = NA, Nx = NA,
  Ny = NA, Nz = NA, plan = NULL, xlim = NULL, ylim = NULL,
  zlim = NULL, vlim = NULL, file.name = NULL, width = 7, height = 7)
}
\description{
display della ct (usa la funzione display.slice)
}
\seealso{
Other display slices: \code{\link{display.slice.all}};
\code{\link{display.slice}};
\code{\link{dumpslices.3d.2.png}}
}

