% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.alpha.beta.beamLUT}
\alias{get.alpha.beta.beamLUT}
\title{Get alpha/beta from beamLUTs (pure-dek)}
\usage{
get.alpha.beta.beamLUT(beamLUTs, beams, x, y, z,
  variables = c("Alpha[Gy^(-1)]", "Beta[Gy^(-2)]"))
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x,y,z}{the coordinates of the 3D array.}

\item{variable}{name vector to override the default name of the variables (c('Alpha[Gy^(-1)]', 'Beta[Gy^(-2)]')).}
}
\value{
a values object.
}
\description{
Get the net alpha and beta Linear-Quadratic parameters distribution (values object) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.bdose.beamLUT}};
  \code{\link{get.beamLUTs}};
  \code{\link{get.dose.beamLUT}};
  \code{\link{get.letd.beamLUT}};
  \code{\link{get.rbe.beamLUT}};
  \code{\link{get.values.for.beam}};
  \code{\link{read.beamLUT}}; \code{\link{write.beamLUT}}
}

