% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.array.values}
\alias{get.array.values}
\title{estrai array della variabile specificata dai values}
\usage{
get.array.values(values, variable)
}
\description{
estrai array della variabile specificata dai values
}
\seealso{
Other Values Manipulation: \code{\link{add.array.values}};
  \code{\link{get.origin}};
  \code{\link{get.subset.values}}; \code{\link{set.origin}}
}

