% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{get.beamLUTs}
\alias{get.beamLUTs}
\title{Get beamLUTs}
\usage{
get.beamLUTs(plan, preallocate = FALSE)
}
\arguments{
\item{plan}{The plan object.}

\item{preallocate}{Preallocate memory (parameter passed to read.beamLUT().)}
}
\description{
Get the evaluated beamLUTs from the plan
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}};
  \code{\link{get.bdose.beamLUT}};
  \code{\link{get.dose.beamLUT}};
  \code{\link{get.letd.beamLUT}};
  \code{\link{get.rbe.beamLUT}};
  \code{\link{get.values.for.beam}};
  \code{\link{read.beamLUT}}; \code{\link{write.beamLUT}}
}

