\name{get.beams}
\alias{get.beams}
\title{Get beams from plan}
\usage{
get.beams(plan, ...)
}
\description{
input serve a selezionare i beams usati come "input" per il
calcolo di default si recuperano quelle in "output". in
teoria sarebbero identici tranne casi particolari per il
posizionamento degli spot (se è stata cambiata CT tra inv e
fwd planning).
}
\seealso{
Other Beams: \code{\link{add.field}};
\code{\link{convert.fluence2mu}};
\code{\link{convert.mu2fluence}};
\code{\link{create.beam}}; \code{\link{create.field}};
\code{\link{display.beamports}};
\code{\link{display.beams}};
\code{\link{get.ct.entrance.point}};
\code{\link{get.isocenter}}; \code{\link{get.rays}};
\code{\link{read.beams.fluka}}; \code{\link{read.beams}};
\code{\link{write.beams}}
}

