% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{get.dektools.env}
\alias{get.dektools.env}
\title{Get Environment Variables}
\usage{
get.dektools.env()
}
\description{
Get Environment Variables
}
\seealso{
Other Environment: \code{\link{check.gate}};
  \code{\link{check.plankit}};
  \code{\link{set.dektools.env}};
  \code{\link{set.lut.folder}};
  \code{\link{setenv.rplanit}}
}

