% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.letd.beamLUT}
\alias{get.letd.beamLUT}
\title{Get dose averaged LET from beamLUTs (pure-dek)}
\usage{
get.letd.beamLUT(beamLUTs, beams, x, y, z,
  variable = "DoseAveragedLET[keV/um]")
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x,y,z}{the coordinates of the 3D array.}

\item{variable}{to override the default name of the variable ('DoseAveragedLET[keV/um]').}
}
\value{
a values object.
}
\description{
Get the net dose averged LET distribution (values object) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}};
  \code{\link{get.bdose.beamLUT}};
  \code{\link{get.beamLUTs}};
  \code{\link{get.dose.beamLUT}};
  \code{\link{get.rbe.beamLUT}};
  \code{\link{get.values.for.beam}};
  \code{\link{read.beamLUT}}; \code{\link{write.beamLUT}}
}

