% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.origin}
\alias{get.origin}
\title{Get origin coordinates}
\usage{
get.origin(values)
}
\arguments{
\item{values}{the values object}
}
\value{
a vector of the coordinates x, y and z.
}
\description{
Get the coordinates of origin corner (corner of voxel [1,1,1]).
}
\seealso{
Other Values Manipulation: \code{\link{add.array.values}};
  \code{\link{get.array.values}};
  \code{\link{get.subset.values}}; \code{\link{set.origin}}
}

