% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescription-functions.R
\name{get.prescription}
\alias{get.prescription}
\title{recupera prescrizione}
\usage{
get.prescription(plan)
}
\description{
recupera prescrizione
}
\seealso{
Other Prescription: \code{\link{check.prescription}},
  \code{\link{create.constraint}}
}
\concept{Prescription}
