% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.rays}
\alias{get.rays}
\title{Get rays from beams}
\usage{
get.rays(beams, s = 4000, unique.rays = FALSE)
}
\arguments{
\item{beams}{the beams data.frame}

\item{s}{distance of the source from the isocenter (it uses only one virtual source).}

\item{unique.rays}{Return only unique rays}
}
\value{
data frame of rays. Each ray has 6 component: a point (X,Y,Z) on the plane passing through the isocenter, and a normalized vector for the direction (xn, yn, zn).
}
\description{
Get rays (normalized vector directions + absolute coordinate) in the reference frame of the CT.
}
\seealso{
Other Beams: \code{\link{add.field}};
  \code{\link{convert.fluence2mu}};
  \code{\link{convert.mu2fluence}};
  \code{\link{create.beam}}; \code{\link{create.field}};
  \code{\link{display.beamports}};
  \code{\link{display.beams}}; \code{\link{get.beams}};
  \code{\link{get.ct.entrance.point}};
  \code{\link{get.isocenter}};
  \code{\link{read.beams.fluka}}; \code{\link{read.beams}};
  \code{\link{write.beams}}
}

