% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{get.rbe.beamLUT}
\alias{get.rbe.beamLUT}
\title{Get RBE from beamLUTs (pure-dek)}
\usage{
get.rbe.beamLUT(beamLUTs, beams, x = NULL, y = NULL, z = NULL,
  variable = "RBE", alphaX, betaX)
}
\arguments{
\item{beamLUTs}{the beamLUTs data frame.}

\item{beams}{the beams data frame.}

\item{x, y, z}{the coordinates of the 3D array.}

\item{variable}{to override the default name of the variable ('RBE').}

\item{alphaX, betaX}{the Linear-Quadratic parameters of the reference radiation (they can be two arrays/vectors)}
}
\value{
A values object (if  x, y and z are specified), otherwise a sparse array.
}
\description{
Get the net RBE distribution (values object or sparse array) from the beamLUTs.
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}},
  \code{\link{get.bdose.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.letd.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{read.beamLUT}}, \code{\link{write.beamLUT}}
}

