\name{get.sparse.array.from.values}
\alias{get.sparse.array.from.values}
\title{recupera matrice sparsa da values}
\usage{
get.sparse.array.from.values(values, variable = "Dose[Gy]", threshold = 0)
}
\description{
accetta values con più di una variabile. la selezione è
fatta sulla prima. (che può coincidere o meno con la
variabile specificata)
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.profile}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values.gate}}; \code{\link{get.values}};
\code{\link{merge.values}}; \code{\link{sanitize.values}}
}

