% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.sparse.array.gate}
\alias{get.sparse.array.gate}
\title{Get sparse array values (Gate)}
\usage{
get.sparse.array.gate(plan.gate)
}
\arguments{
\item{plan.gate}{The plan object.}
}
\description{
Get the sparse array evaluated for a plan (Gate).
Note: the plan has to be evaluated via run.gate.forward() with the option evaluate.sparse.arrays=TRUE.
}
\seealso{
Other Values: \code{\link{combine.values}};
  \code{\link{create.values}};
  \code{\link{dataframe.from.values}};
  \code{\link{get.values.dataframe}};
  \code{\link{get.values.gate.plan}};
  \code{\link{get.values.plankit.plan}};
  \code{\link{get.values}}; \code{\link{sanitize.values}};
  \code{\link{sparse.array.from.values}};
  \code{\link{values.from.sparse.array}}
}

