% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.subset.vois}
\alias{get.subset.vois}
\title{Get a subset of a  \code{vois} object.}
\usage{
get.subset.vois(vois, xlim = c(-Inf, Inf), ylim = c(-Inf, Inf),
  zlim = c(-Inf, Inf))
}
\arguments{
\item{xlim,ylim,zlim}{The ranges of the subset coordinates.}

\item{values}{the \code{vois} object}
}
\description{
Get a subset of a  \code{vois} object.
}
\seealso{
Other VOIs: \code{\link{create.vois}};
  \code{\link{get.voi.logical}};
  \code{\link{get.voiindex}}; \code{\link{get.vois}};
  \code{\link{is.in.voi}}
}

