% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.values.dataframe}
\alias{get.values.dataframe}
\title{recupera matrice di valori in formato dataframe}
\usage{
\method{get.values}{dataframe}(plan)
}
\description{
Obsoleto
}
\seealso{
Other Values: \code{\link{combine.values}};
  \code{\link{create.values}};
  \code{\link{dataframe.from.values}};
  \code{\link{get.sparse.array.gate}};
  \code{\link{get.values.gate.plan}};
  \code{\link{get.values.list}};
  \code{\link{get.values.plankit.plan}};
  \code{\link{get.values}}; \code{\link{sanitize.values}};
  \code{\link{sparse.array.from.values}};
  \code{\link{values.from.sparse.array}}
}

