\name{get.values.dataframe}
\alias{get.values.dataframe}
\title{recupera matrice di valori in formato dataframe}
\usage{
get.values.dataframe(plan)
}
\description{
recupera matrice di valori in formato dataframe
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.profile}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values}}; \code{\link{sanitize.values}}
}

