% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.values.gate.plan}
\alias{get.values.gate.plan}
\title{Get values object (Gate)}
\usage{
\method{get.values}{gate.plan}(plan, center.coordinates = NULL)
}
\arguments{
\item{center.coordinates}{The coordinates of the center of the values 3D distribution. If it is NULL, the center of the CT is assumed.}

\item{plan.gate}{The plan object.}
}
\description{
Get the values object associated to a plan (Gate).
}
\seealso{
Other Values: \code{\link{combine.values}},
  \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{merge.values}},
  \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{threshold.values}},
  \code{\link{values.from.sparse.array}}
}
\concept{Values}
