\name{get.values.plankit.plan}
\alias{get.values.plankit.plan}
\title{Get values object (PlanKIT)}
\usage{
\method{get.values}{plankit.plan}(plan)
}
\arguments{
  \item{plan}{The plan object.}
}
\description{
Get the values matrices for a plan (PlanKIT).
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.sparse.array.gate}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values.gate.plan}};
\code{\link{get.values}}; \code{\link{merge.values}};
\code{\link{sanitize.values}}
}

