% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.values.plankit.plan}
\alias{get.values.plankit.plan}
\title{Get values object (PlanKIT)}
\usage{
\method{get.values}{plankit.plan}(plan, input = FALSE)
}
\arguments{
\item{plan}{The plan object.}

\item{input}{Get the input values (if exits).}
}
\description{
Get the values object associated to a plan (PlanKIT).
}
\seealso{
Other Values: \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}}, \code{\link{get.values}},
  \code{\link{merge.values}},
  \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{threshold.values}},
  \code{\link{values.from.sparse.array}}
}
