\name{get.values}
\alias{get.values}
\title{recupera matrice di valori}
\usage{
get.values(plan)
}
\description{
recupera matrice di valori
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.profile}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{sanitize.values}}
}

