\name{get.values}
\alias{get.values}
\title{Get values object (PlanKIT)}
\usage{
get.values(plan)
}
\arguments{
  \item{plan}{The plan object.}
}
\description{
Get the values matrices for a plan (PlanKIT).
}
\seealso{
Other Values: \code{\link{create.values}};
\code{\link{get.dataframe.from.values}};
\code{\link{get.sparse.array.from.values}};
\code{\link{get.sparse.array.gate}};
\code{\link{get.values.dataframe}};
\code{\link{get.values.from.sparse.array}};
\code{\link{get.values.gate}}; \code{\link{merge.values}};
\code{\link{sanitize.values}}
}

