% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.values}
\alias{get.values}
\title{Get values object}
\usage{
get.values(plan, ...)
}
\arguments{
\item{plan}{The plan object.}
}
\description{
Get the values matrices for a plan.
}
\seealso{
Other Values: \code{\link{combine.values}},
  \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{sanitize.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{threshold.values}},
  \code{\link{values.from.sparse.array}}
}
