% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.values_at.points}
\alias{get.values_at.points}
\title{get values at coordinates}
\usage{
get.values_at.points(values, x, y, z, variable = "Dose[Gy]")
}
\arguments{
\item{values}{Values object}

\item{x,y,z}{coordinates of a 3D point on which to evaluate the values (they can be vectors).}
}
\value{
a vector of values.
}
\description{
get values at coordinates
}
\seealso{
Other ValuesUtilities: \code{\link{identify.slice}};
  \code{\link{interpolate.values}}; \code{\link{meshgrid}}
}

