\name{get.voi.logical}
\alias{get.voi.logical}
\title{Evaluate VOI logical matrix}
\usage{
get.voi.logical(vois, voi = NULL)
}
\arguments{
  \item{vois}{the \code{vois} object}

  \item{voi}{the name of the VOI}
}
\value{
boolean array representing the volumetric occupancy of a
specific VOI.
}
\description{
Evaluate a boolean 3D array representing the volumetric
occupancy of a specific VOI.
}
\seealso{
Other VOIs: \code{\link{get.voiindex}};
\code{\link{get.vois}}
}

