\name{get.voxel.index}
\alias{get.voxel.index}
\title{Get voxel index from coordinates}
\usage{
get.voxel.index(x, y, z, Xv, Yv, Zv)
}
\arguments{
  \item{x,y,z}{coordinate vectors.}

  \item{Xv,Yv,Zv}{coordinates of the voxels.}
}
\value{
A vector.
}
\description{
Get voxel index from coordinates
}
\seealso{
Other Values, Utilities: \code{\link{create.intervals}};
\code{\link{generate.values.from.events}};
\code{\link{get.coordinates}}; \code{\link{get.profile}}
}

