% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{get.voxel.index}
\alias{get.voxel.index}
\title{Get voxel index from coordinates}
\usage{
get.voxel.index(x, y, z, Xv, Yv, Zv)
}
\arguments{
\item{x, y, z}{coordinate vectors.}

\item{Xv, Yv, Zv}{coordinates of the voxels.}
}
\value{
A vector.
}
\description{
Get voxel index from coordinates
}
\seealso{
Other Values, Values Utilities: \code{\link{create.intervals}},
  \code{\link{generate.values.from.events}},
  \code{\link{get.coordinates}}
}
