\name{get.zBP.fun}
\alias{get.zBP.fun}
\title{Legge file zBP vs. E e ritorna una funzione d'interpolazione}
\usage{
get.zBP.fun(lut.name, EBP = FALSE)
}
\description{
Legge file zBP vs. E e ritorna una funzione
d'interpolazione
}
\seealso{
Other LUT: \code{\link{lut.add.alpha}};
\code{\link{lut.add.beta}};
\code{\link{lut.add.rbe.alpha}};
\code{\link{lut.radial.integral}}; \code{\link{read.lut}};
\code{\link{write.lut}}
}

