% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{identify.slice}
\alias{identify.slice}
\title{identifica slice corrispondente alla coordinata}
\usage{
\method{identify}{slice}(x, xvec)
}
\description{
identifica slice corrispondente alla coordinata
}
\seealso{
Other ValuesUtilities: \code{\link{get.values_at.points}},
  \code{\link{interpolate.values}}, \code{\link{meshgrid}}
}
\concept{ValuesUtilities}
