% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{install.gate}
\alias{install.gate}
\title{Install Gate}
\usage{
install.gate()
}
\description{
Install Gate
}
\seealso{
Other Install: \code{\link{install.plankit}},
  \code{\link{install.puredek}},
  \code{\link{install.survival}}
}
\concept{Install}
