% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{install.plankit}
\alias{install.plankit}
\title{Install Plankit (obsolete)}
\usage{
install.plankit()
}
\description{
Install Plankit (obsolete)
}
\seealso{
Other Install: \code{\link{install.gate}};
  \code{\link{install.puredek}};
  \code{\link{install.survival}}
}

