% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{install.puredek}
\alias{install.puredek}
\title{Install Pure-dek}
\usage{
install.puredek()
}
\description{
Install Pure-dek
}
\seealso{
Other Install: \code{\link{install.gate}};
  \code{\link{install.plankit}};
  \code{\link{install.survival}}
}

