% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{install.survival}
\alias{install.survival}
\title{Install Survival}
\usage{
install.survival()
}
\description{
Install Survival
}
\seealso{
Other Install: \code{\link{install.gate}},
  \code{\link{install.plankit}},
  \code{\link{install.puredek}}
}
\concept{Install}
