% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{interpolate.values}
\alias{interpolate.values}
\title{Tri-linear interpolation}
\usage{
interpolate.values(values, x, y, z)
}
\arguments{
\item{values}{Values object}

\item{x, y, z}{cartesian coordinate on which evaluate the interpolation.}
}
\value{
a values object (with elements at coordinates x,y,z).
}
\description{
Tri-linear interpolation
}
\seealso{
Other ValuesUtilities: \code{\link{get.values_at.points}},
  \code{\link{identify.slice}}, \code{\link{meshgrid}}
}
\concept{ValuesUtilities}
