% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lut.add.alpha}
\alias{lut.add.alpha}
\title{Add alpha values to lUT}
\usage{
lut.add.alpha(lut, tissue.number = 1, tissue = NULL)
}
\arguments{
\item{lut}{the lut object}

\item{tissue.number}{the number of the tissue stored in the LUT (start from 1)}
}
\description{
Add variable Alpha[Gy^(-1)] = LethAlphaPerEvent / DosePerEvent
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
  \code{\link{lut.add.beta}};
  \code{\link{lut.add.rbe.alpha}};
  \code{\link{lut.radial.integral}};
  \code{\link{read.lut}}; \code{\link{write.lut}}
}

