% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{lut.add.alpha}
\alias{lut.add.alpha}
\title{Add alpha values to lUT}
\usage{
lut.add.alpha(lut, tissue.number = 1, tissue = NULL)
}
\arguments{
\item{lut}{the lut object}

\item{tissue.number}{the number of the tissue stored in the LUT (start from 1)}
}
\description{
Add variable Alpha[Gy^(-1)] = LethAlphaPerEvent / DosePerEvent
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}},
  \code{\link{lut.add.beta}},
  \code{\link{lut.add.rbe.alpha}},
  \code{\link{lut.radial.integral}},
  \code{\link{read.lut}}, \code{\link{write.lut.4d}},
  \code{\link{write.lut.meta}},
  \code{\link{write.lut.optics}},
  \code{\link{write.lut.phasespace}},
  \code{\link{write.lut.zBP}}, \code{\link{write.lut}}
}
\concept{LUT}
