% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{lut.add.beta}
\alias{lut.add.beta}
\title{aggiungi beta alla lut}
\usage{
lut.add.beta(lut)
}
\description{
la formula: LethAlphaPerEvent / DosePerEvent
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}},
  \code{\link{lut.add.alpha}},
  \code{\link{lut.add.rbe.alpha}},
  \code{\link{lut.radial.integral}},
  \code{\link{read.lut}}, \code{\link{write.lut.4d}},
  \code{\link{write.lut.meta}},
  \code{\link{write.lut.optics}},
  \code{\link{write.lut.phasespace}},
  \code{\link{write.lut.zBP}}, \code{\link{write.lut}}
}
\concept{LUT}
