% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{lut.add.rbe.alpha}
\alias{lut.add.rbe.alpha}
\title{Aggiunge RBE.alpha alla LUT}
\usage{
lut.add.rbe.alpha(lut, alphaX)
}
\description{
Aggiunge RBE.alpha alla LUT
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}},
  \code{\link{lut.add.alpha}}, \code{\link{lut.add.beta}},
  \code{\link{lut.radial.integral}},
  \code{\link{read.lut}}, \code{\link{write.lut.4d}},
  \code{\link{write.lut.meta}},
  \code{\link{write.lut.optics}},
  \code{\link{write.lut.phasespace}},
  \code{\link{write.lut.zBP}}, \code{\link{write.lut}}
}

