% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{lut.radial.integral}
\alias{lut.radial.integral}
\title{integra la lut radiale}
\usage{
lut.radial.integral(lut, r.cut = Inf)
}
\description{
integra la lut radiale
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
  \code{\link{lut.add.alpha}}; \code{\link{lut.add.beta}};
  \code{\link{lut.add.rbe.alpha}}; \code{\link{read.lut}};
  \code{\link{write.lut.4d}}; \code{\link{write.lut.meta}};
  \code{\link{write.lut.optics}};
  \code{\link{write.lut.phasespace}};
  \code{\link{write.lut.zBP}}; \code{\link{write.lut}}
}

