% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{meshgrid}
\alias{meshgrid}
\title{Create a meshgrid}
\usage{
meshgrid(x, y, z)
}
\arguments{
\item{x, y, z}{Coordinates vector.}
}
\value{
A list with X, Y and Z 3D arrays
}
\description{
Generate X, Y and Z matrices for three-dimensional arrays starting from x, y and z vectors. The function is similar to the Matlab function "meshgrid".
}
\seealso{
Other ValuesUtilities: \code{\link{get.values_at.points}},
  \code{\link{identify.slice}},
  \code{\link{interpolate.values}}
}
\concept{ValuesUtilities}
