% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{mkm.lower}
\alias{mkm.lower}
\title{MKM lower limits}
\format{\preformatted{ num [1:4] 0.1 0.02 1 0.1
}}
\usage{
mkm.lower
}
\description{
MKM lower limits
}
\keyword{datasets}

