% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{mkm.upper}
\alias{mkm.upper}
\title{MKM upper limits}
\format{An object of class \code{numeric} of length 4.}
\usage{
mkm.upper
}
\description{
MKM upper limits
}
\keyword{datasets}
