% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading-writing-functions.R
\name{read.1d}
\alias{read.1d}
\title{Read 1D array}
\usage{
read.1d(file.name)
}
\arguments{
\item{file.name}{the file name}
}
\value{
a dataframe
}
\description{
Read a file "*.1d" (PlanKIT format) and return a dataframe.
}
\seealso{
Other R/W Arrays: \code{\link{read.3d.dicom}};
  \code{\link{read.3d.hdr}}; \code{\link{read.3d}}
}

