% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{read.beamLUT}
\alias{read.beamLUT}
\title{Read a beamLUT (pure-dek)}
\usage{
read.beamLUT(beamLUT.file, Nx = NULL, Ny = NULL, Nz = NULL,
  dose.threshold = 0)
}
\arguments{
\item{beamLUT.file}{the name of the beamLUT file.}

\item{Nx, Ny, Nz}{optional dimensions of the target 3D array (to evaluate an absolute voxelID)}

\item{dose.threshold}{additional dose threshold to filter elements with dose <= dose.threshold.}
}
\value{
a data.table of the beamLUT
}
\description{
Read a beamLUT (pure-dek)
}
\seealso{
Other BeamLUT: \code{\link{combine.beamLUTs}},
  \code{\link{get.alpha.beta.beamLUT}},
  \code{\link{get.bdose.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.letd.beamLUT}},
  \code{\link{get.rbe.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{read.beamLUTs}}, \code{\link{write.beamLUT}}
}
\concept{BeamLUT}
