% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{read.beamLUT}
\alias{read.beamLUT}
\title{Read the beamLUTs (pure-dek)}
\usage{
read.beamLUT(beamLUT.name, Nx, Ny, Nz, dose.threshold = 0,
  preallocate = FALSE)
}
\arguments{
\item{beamLUT.name}{the prefix of the beamLUT files. The complete filenames are expected to be: <beamLUT.name>_<beamIndex>.beamLUT}

\item{Nx, Ny, Nz}{the dimensions of the target 3D array (to evaluate an absolute voxelID)}

\item{dose.threshold}{additional dose threshold to filter elements with dose <= dose.threshold.}
}
\description{
Read the beamLUTs (pure-dek)
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}},
  \code{\link{get.bdose.beamLUT}},
  \code{\link{get.beamLUTs}},
  \code{\link{get.dose.beamLUT}},
  \code{\link{get.letd.beamLUT}},
  \code{\link{get.rbe.beamLUT}},
  \code{\link{get.values.for.beam}},
  \code{\link{write.beamLUT}}
}

