% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.beams}
\alias{read.beams}
\title{Read beams from file (PlanKIT)}
\usage{
read.beams(beams.file)
}
\arguments{
\item{beams.file}{The file name.}
}
\value{
a beams object.
}
\description{
Read RtIonPlan file (Plankit *.beams format) and return a beams object.
}
\seealso{
Other Beams: \code{\link{add.field}};
  \code{\link{convert.fluence2mu}};
  \code{\link{convert.mu2fluence}};
  \code{\link{create.beam}}; \code{\link{create.field}};
  \code{\link{display.beamports}};
  \code{\link{display.beams}}; \code{\link{get.beams}};
  \code{\link{get.ct.entrance.point}};
  \code{\link{get.isocenter}}; \code{\link{get.rays}};
  \code{\link{read.beams.fluka}}; \code{\link{write.beams}}
}

