% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-functions.R
\name{read.lut}
\alias{read.lut}
\title{Read LUT (4d)}
\usage{
read.lut(lut.name, dataframe = FALSE)
}
\arguments{
\item{lut.name}{the name (prefix of the file name) of the lut (i.e. not including the file name extension.)}

\item{dataframe}{return a dataframe.}
}
\description{
Read pencil-LUT.
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}},
  \code{\link{lut.add.alpha}}, \code{\link{lut.add.beta}},
  \code{\link{lut.add.rbe.alpha}},
  \code{\link{lut.radial.integral}},
  \code{\link{write.lut.4d}}, \code{\link{write.lut.meta}},
  \code{\link{write.lut.optics}},
  \code{\link{write.lut.phasespace}},
  \code{\link{write.lut.zBP}}, \code{\link{write.lut}}
}

