\name{read.lut}
\alias{read.lut}
\title{Legge LUT}
\usage{
read.lut(lut.name, dataframe = FALSE)
}
\description{
Legge pencil-LUT da un file .4d e restituisce un oggetto
lut (opzionalmente un dataframe)
}
\seealso{
Other LUT: \code{\link{get.zBP.fun}};
\code{\link{lut.add.alpha}}; \code{\link{lut.add.beta}};
\code{\link{lut.add.rbe.alpha}};
\code{\link{lut.radial.integral}}; \code{\link{write.lut}}
}

