% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plankit-functions.R
\name{read.plan}
\alias{read.plan}
\title{Read Plan}
\usage{
read.plan(name)
}
\arguments{
\item{name}{a string name of the plan. It can be a vector of names. In the latter case a list of plan objects is returned.}
}
\value{
the Plan object (or a list of Plan objects)
}
\description{
Read a Plan object from disk. The origin folder is derived from the plan name.
}
\seealso{
Other Plan: \code{\link{create.plan}};
  \code{\link{save.plan}}
}

