% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{remove.atomic.mass}
\alias{remove.atomic.mass}
\title{Remove atomic mass number}
\usage{
remove.atomic.mass(ions)
}
\arguments{
\item{ions}{A character vector containing the ion names (including the atomic mass number specification).}
}
\value{
A character vector containing the ion names (without the atomic mass number specification).
}
\description{
Returns the name of the ion/atom removing its atomic mass number specification (e.g. it returns "C" from "12C").
Note: the available range is from 1H to 20 Ne.
}
\seealso{
Other Particles: \code{\link{a.from.ion}},
  \code{\link{ion.from.z}}, \code{\link{z.from.ion}}
}
