% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sanitize.filename}
\alias{sanitize.filename}
\title{Sanitize file name}
\usage{
sanitize.filename(filename)
}
\arguments{
\item{filename}{the file name}
}
\value{
sanitized file name
}
\description{
Sanitize file name
}
\seealso{
Other Utilities: \code{\link{evaluate.bands}};
  \code{\link{get.filepath}}
}

