% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{sanitize.values}
\alias{sanitize.values}
\title{Set not defined or not "physical" values to NA}
\usage{
sanitize.values(values)
}
\arguments{
\item{values}{a values object}
}
\value{
a values object with NA where the radiobiological quantities are not defined (e.g. where the dose is zero)
}
\description{
Set not defined or not "physical" values to NA
}
\seealso{
Other Values: \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{merge.values}},
  \code{\link{sparse.array.from.values}},
  \code{\link{threshold.values}},
  \code{\link{values.from.sparse.array}}
}
