% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{set.origin}
\alias{set.origin}
\title{Set origin coordinates}
\usage{
set.origin(values, coordinates)
}
\arguments{
\item{coordinates}{a vector of the coordinates x, y and z.}
}
\value{
the translated values object
}
\description{
Set the coordinates of origin corner (corner of voxel [1,1,1]).
}
\seealso{
Other Values Manipulation: \code{\link{add.array.values}},
  \code{\link{get.array.values}}, \code{\link{get.origin}},
  \code{\link{get.subset.values}}
}
