% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-functions.R
\name{setenv.rplanit}
\alias{setenv.rplanit}
\title{Set the system environment variables}
\usage{
setenv.rplanit()
}
\description{
Set the esystem environment variables for Plankit, Gate and Survival.
}
\seealso{
Other Environment: \code{\link{check.gate}},
  \code{\link{check.plankit}},
  \code{\link{get.dektools.env}},
  \code{\link{set.dektools.env}},
  \code{\link{set.lut.folder}}
}

