% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-functions.R
\name{sparse.array.from.values}
\alias{sparse.array.from.values}
\title{get sparse format data from a values object}
\usage{
sparse.array.from.values(values, variables = NULL, threshold = 0)
}
\arguments{
\item{values}{the values object}

\item{variables}{a vector of variable names. If it is not specified, all variabels stored in values are extracted in the sparse matrix.}

\item{threshold}{applay a threshold selection on the first variable stored in values (usually 'Dose[Gy]')}
}
\description{
get sparse format data from a values object
}
\seealso{
Other Values: \code{\link{create.values}},
  \code{\link{dataframe.from.values}},
  \code{\link{get.sparse.array.gate}},
  \code{\link{get.values.dataframe}},
  \code{\link{get.values.gate.plan}},
  \code{\link{get.values.list}},
  \code{\link{get.values.plankit.plan}},
  \code{\link{get.values}}, \code{\link{merge.values}},
  \code{\link{sanitize.values}},
  \code{\link{threshold.values}},
  \code{\link{values.from.sparse.array}}
}
