% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{srim.es.max}
\alias{srim.es.max}
\title{srim.limits}
\format{\preformatted{ Named num [1:10] 10000 2500 1428 1111 909 ...
 - attr(*, "names")= chr [1:10] "H" "He" "Li" "Be" ...
}}
\usage{
srim.es.max
}
\description{
srim.limits
}
\keyword{datasets}

