% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological-functions.R
\docType{data}
\name{srim.es.min}
\alias{srim.es.min}
\title{srim.limits}
\format{An object of class \code{numeric} of length 10.}
\usage{
srim.es.min
}
\description{
srim.limits
}
\keyword{datasets}
