% Generated by roxygen2 (4.0.2): do not edit by hand
\name{volume.contours}
\alias{volume.contours}
\title{Evaluate volume from contours}
\usage{
volume.contours(contours, voi = NULL)
}
\arguments{
\item{contours}{the contours dataframe}

\item{voi}{the voi(s). If it is not specified, the volume of all VOIs is evaluated (optional)}
}
\description{
Evaluate the volume defined by its contours.
}
\seealso{
Other Contours: \code{\link{get.contours}};
  \code{\link{sanitize.contours}}
}

