% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.beamLUT}
\alias{write.beamLUT}
\title{scrive le matrici V_{b,v} (beamLUT)}
\usage{
write.beamLUT(plan, threshold = 0, threshold.variable = "Dose[Gy]",
  variable = "Dose[Gy]", file.name)
}
\description{
nota: è usata la "threshold.variable" (normalmente la dose) per fissare la soglia
le lut sono normalizzate per singolo primario
le lut sono calcolate su tutto il volume usato in "plan"
}
\seealso{
Other BeamLUT: \code{\link{get.alpha.beta.beamLUT}};
  \code{\link{get.bdose.beamLUT}};
  \code{\link{get.beamLUTs}};
  \code{\link{get.dose.beamLUT}};
  \code{\link{get.letd.beamLUT}};
  \code{\link{get.rbe.beamLUT}};
  \code{\link{get.values.for.beam}};
  \code{\link{read.beamLUT}}
}

